--- vnexpress_data_warehouse/spiders/TonghopData_Cacbenh.py
@@ -1,1307 +1,0 @@
-import scrapy
-# import os
-from vnexpress_data_warehouse.items import VnexpressDataWarehouseItem 
-class TonghopData_CacbenhSpider(scrapy.Spider):
-    name = "TonghopData_Cacbenh"
-    allowed_domains = ["vnexpress.net"]
-    start_urls = [
-        'https://vnexpress.net/nhan-giap-tien-trien-ung-thu-4785610.html',
-        'https://vnexpress.net/cuong-giap-loi-mat-do-benh-basedow-4784710.html',
-        'https://vnexpress.net/tuong-mac-benh-tam-ly-hoa-cuong-giap-4784388.html',
-        'https://vnexpress.net/buou-co-to-gay-kho-tho-4783549.html',
-        'https://vnexpress.net/bach-cau-giam-manh-do-tac-dung-phu-cua-thuoc-chua-cuong-giap-4759724.html',
-        'https://vnexpress.net/co-phinh-to-boi-nang-giap-4741323.html',
-        'https://vnexpress.net/kham-suc-khoe-dinh-ky-phat-hien-ung-thu-tuyen-giap-4726805.html',
-        'https://vnexpress.net/bat-ngo-phat-hien-ung-thu-khi-kham-suc-khoe-4725670.html',
-        'https://vnexpress.net/u-tuyen-giap-lanh-tinh-hoa-ung-thu-sau-10-nam-4702762.html',
-        'https://vnexpress.net/phat-hien-cuong-giap-tu-dau-hieu-kho-tho-tim-dap-nhanh-4700814.html',
-        'https://vnexpress.net/suy-tim-do-cuong-giap-ma-khong-biet-4690458.html',
-        'https://vnexpress.net/tuong-benh-tuoi-gia-hoa-cuong-giap-bien-chung-tim-4688789.html',
-        'https://vnexpress.net/dot-ngot-noi-ngong-do-suy-giap-4685997.html',
-        'https://vnexpress.net/kho-tho-boi-buou-co-khong-lo-4676134.html',
-        'https://vnexpress.net/cuong-giap-khien-nguoi-dan-ong-liet-hai-chan-4669409.html',
-        'https://vnexpress.net/tuong-dot-quy-hoa-u-tuyen-thuong-than-4653156.html',
-        'https://vnexpress.net/suy-tim-do-giam-hormone-tuyen-giap-4644868.html',
-        'https://vnexpress.net/tuong-sut-can-hoa-cuong-giap-nang-4638161.html',
-        'https://vnexpress.net/phat-hien-benh-tuyen-giap-khi-kham-tim-4637591.html',
-        'https://vnexpress.net/suy-giap-sau-10-nam-cat-tuyen-giap-4633800.html',
-        'https://vnexpress.net/khoi-u-moc-chan-xam-lan-tuyen-giap-4622278.html',
-        'https://vnexpress.net/roi-loan-nhip-tim-do-quen-uong-thuoc-tri-benh-tuyen-giap-4609253.html',
-        'https://vnexpress.net/cuu-co-gai-19-tuoi-khoi-suy-tim-do-cuong-giap-4608376.html',
-        'https://vnexpress.net/loan-nhip-tim-khien-nguoi-benh-cuong-giap-nguy-kich-4505434.html',
-        'https://vnexpress.net/mo-buou-giap-ac-tinh-cho-benh-nhan-beo-phi-4496844.html',
-        'https://vnexpress.net/cuong-giap-suy-gan-nang-o-benh-nhan-tieu-duong-4467644.html',
-        "https://vnexpress.net/co-nen-dung-don-thuoc-chua-mun-trung-ca-cua-nguoi-khac-4788048.html",
-        "https://vnexpress.net/5-lam-tuong-ve-mun-4786114.html",
-        "https://vnexpress.net/nhieu-nguoi-nhiem-nam-viem-da-mua-mua-4785838.html",
-        "https://vnexpress.net/do-mo-hoi-uot-ao-du-o-phong-dieu-hoa-la-benh-gi-4785376.html",
-        "https://vnexpress.net/dau-hieu-nhiem-giun-san-tren-da-4784809.html",
-        "https://vnexpress.net/cach-nao-dieu-tri-nhanh-mun-trung-ca-4784321.html",
-        "https://vnexpress.net/6-rau-cu-tre-hoa-da-4783976.html",
-        "https://vnexpress.net/7-thoi-quen-hang-ngay-khien-da-lao-hoa-som-4783481.html",
-        "https://vnexpress.net/nhung-vung-da-thuong-bi-bo-quen-thoa-khi-kem-chong-nang-4782531.html",
-        "https://vnexpress.net/au-trung-di-chuyen-ngoan-ngoeo-duoi-da-chang-trai-4782410.html",
-        "https://vnexpress.net/bien-chung-do-tu-dieu-tri-mun-4781466.html",
-        "https://vnexpress.net/loi-ich-khi-tam-nuoc-lanh-4781358.html",
-        "https://vnexpress.net/ly-do-khong-nen-de-toc-uot-di-ngu-4779573.html",
-        "https://vnexpress.net/10-nam-chua-mun-trung-ca-4779178.html",
-        "https://vnexpress.net/4-chuc-nang-cua-da-4778428.html",
-        "https://vnexpress.net/dieu-tri-vay-nen-bang-thuoc-sinh-hoc-4777793.html",
-        "https://vnexpress.net/meo-xoa-sach-mun-lung-4777309.html",
-        "https://vnexpress.net/bong-doc-kien-ba-khoang-vao-mua-mua-4776963.html",
-        "https://vnexpress.net/bien-chung-nguy-hiem-cua-benh-vay-nen-4776531.html",
-        "https://vnexpress.net/nguyen-nhan-gay-tang-tiet-mo-hoi-4774310.html",
-        "https://vnexpress.net/seo-hinh-thanh-nhu-the-nao-4773985.html",
-        "https://vnexpress.net/cach-hap-thu-vitamin-d-an-toan-khi-phoi-nang-4773588.html",
-        "https://vnexpress.net/ruou-anh-huong-den-benh-vay-nen-the-nao-4773188.html",
-        "https://vnexpress.net/lam-the-nao-het-mun-dau-den-4771026.html",
-        "https://vnexpress.net/10-nguyen-nhan-gay-mun-trung-ca-4770174.html",
-        "https://vnexpress.net/ai-de-bi-tan-nhang-4769590.html",
-        "https://vnexpress.net/khi-nao-can-dieu-tri-mun-4769519.html",
-        "https://vnexpress.net/kem-chong-nang-vat-ly-khac-kem-hoa-hoc-the-nao-4768199.html",
-        "https://vnexpress.net/co-the-chua-khoi-vay-nen-khong-4767757.html",
-        "https://vnexpress.net/cach-dung-kem-chong-nang-dang-xit-hieu-qua-4766389.html",
-        "https://vnexpress.net/cach-cham-soc-da-vay-nen-4763655.html",
-        "https://vnexpress.net/bi-viem-da-di-ung-nen-an-gi-4762341.html",
-        "https://vnexpress.net/an-gi-giup-da-hoi-phuc-sau-chay-nang-4762308.html",
-        "https://vnexpress.net/me-day-4761817.html",
-        "https://vnexpress.net/chan-doan-dieu-tri-not-ruoi-ung-thu-the-nao-4760602.html",
-        "https://vnexpress.net/tiem-huyet-tuong-giau-tieu-cau-chua-rung-toc-the-nao-4760155.html",
-        "https://vnexpress.net/nhung-bien-chung-da-thuong-gap-do-my-pham-4759687.html",
-        "https://vnexpress.net/meo-chong-nang-khi-di-may-bay-4758793.html",
-        "https://vnexpress.net/viem-da-co-dia-benh-de-tai-phat-khi-giao-mua-4758194.html",
-        "https://vnexpress.net/toc-rung-bao-nhieu-soi-moi-dang-lo-4757837.html",
-        "https://vnexpress.net/bon-sai-lam-khi-dieu-tri-ung-thu-da-4757393.html",
-        "https://vnexpress.net/hoa-chat-tay-rua-gay-kich-ung-viem-da-to-dia-4756504.html",
-        "https://vnexpress.net/nguoi-tre-tim-cach-chua-rung-toc-4753686.html",
-        "https://vnexpress.net/9-cach-giup-toc-nhanh-moc-4748645.html",
-        "https://vnexpress.net/trieu-chung-nao-nhan-biet-vay-nen-4751024.html",
-        "https://vnexpress.net/co-nen-trang-diem-ngay-nang-nong-4750494.html",
-        "https://vnexpress.net/laser-tri-nam-da-nhu-the-nao-4749624.html",
-        "https://vnexpress.net/hai-kieu-rung-toc-o-phu-nu-4748650.html",
-        "https://vnexpress.net/nhung-van-de-da-thuong-gap-vao-mua-he-4748238.html",
-        "https://vnexpress.net/noi-me-day-boi-di-ung-thuc-an-4747518.html",
-        "https://vnexpress.net/co-kieng-tiep-xuc-khi-mac-benh-zona-4746696.html",
-        "https://vnexpress.net/phan-biet-trieu-chung-zona-va-kien-ba-khoang-tren-da-4746377.html",
-        "https://vnexpress.net/co-nen-nho-toc-bac-toc-sau-4745810.html",
-        "https://vnexpress.net/nang-nong-hai-da-the-nao-4745392.html",
-        "https://vnexpress.net/mac-quan-ao-moi-chua-giat-co-lay-benh-4744885.html",
-        "https://vnexpress.net/soc-xuat-hien-o-mong-tay-la-benh-gi-4744780.html",
-        "https://vnexpress.net/tam-rua-bang-nuoc-am-co-tot-cho-da-4744345.html",
-        "https://vnexpress.net/nguy-hiem-khi-chua-zona-bang-meo-dan-gian-4743822.html",
-        "https://vnexpress.net/8-mon-giau-vitamin-giup-da-dep-toc-khoe-4743423.html",
-        "https://vnexpress.net/tai-sao-tuoi-thieu-nien-da-bac-toc-som-4743389.html",
-        "https://vnexpress.net/6-van-de-suc-khoe-bieu-hien-qua-mai-toc-4742415.html",
-        "https://vnexpress.net/6-benh-gay-rung-toc-4739662.html",
-        'https://vnexpress.net/bui-toc-dai-20-cm-lam-tac-ruot-be-gai-4784006.html',
-        'https://vnexpress.net/nhung-mon-an-uong-giup-giam-o-nong-4782973.html',
-        'https://vnexpress.net/nhung-mon-an-thuc-uong-giup-nhuan-trang-4782543.html',
-        'https://vnexpress.net/nhung-loai-nuoc-nen-uong-va-nen-tranh-khi-o-nong-4782418.html',
-        'https://vnexpress.net/lam-gi-ngay-sau-khi-non-4782206.html',
-        'https://vnexpress.net/suc-khoe/cac-benh/benh-tieu-hoa/benh-duong-tieu-hoa-p2',
-        'https://vnexpress.net/an-banh-ngot-buoi-sang-gay-hai-tieu-hoa-the-nao-4786798.html',
-        'https://vnexpress.net/6-do-uong-giup-nhuan-trang-4786079.html',
-        'https://vnexpress.net/7-loai-tra-giup-giam-day-hoi-4785622.html',
-        'https://vnexpress.net/cac-trieu-chung-tieu-hoa-canh-bao-benh-4785101.html',
-        'https://vnexpress.net/nhung-mon-an-uong-co-the-gay-hoi-chung-ruot-kich-thich-4784879.html',
-        'https://vnexpress.net/suy-kiet-do-viem-loet-dai-trang-boi-nhiem-virus-4784665.html',
-        'https://vnexpress.net/nguyen-nhan-dau-bung-sau-an-trai-cay-4784293.html',
-        'https://vnexpress.net/dieu-tri-nuot-nghen-do-trao-nguoc-da-day-4778285.html',
-        'https://vnexpress.net/9-thuc-pham-can-tranh-an-khi-kho-tieu-4777722.html',
-        'https://vnexpress.net/benh-celiac-4776998.html',
-        'https://vnexpress.net/trao-nguoc-da-day-thuc-quan-do-thoi-quen-sai-lam-4777225.html',
-        'https://vnexpress.net/stress-gay-roi-loan-tieu-hoa-nhu-the-nao-4776864.html',
-        'https://vnexpress.net/ai-nen-su-dung-men-tieu-hoa-4776378.html',
-        'https://vnexpress.net/8-mon-an-uong-nguoi-dau-da-day-nen-han-che-4776408.html',
-        'https://vnexpress.net/loi-ich-khi-uong-nuoc-gung-sau-bua-an-4775647.html',
-        'https://vnexpress.net/liet-da-day-4774716.html',
-        'https://vnexpress.net/nhung-loai-trai-cay-nen-han-che-an-luc-doi-4773882.html',
-        'https://vnexpress.net/an-uong-gi-giup-giam-dau-da-day-nhanh-4773204.html',
-        'https://vnexpress.net/hep-hau-mon-bien-chung-nang-sau-mo-tri-4771995.html',
-        'https://vnexpress.net/nguyen-nhan-hep-thuc-quan-4770946.html',
-        'https://vnexpress.net/dung-thuoc-sai-cach-hai-da-day-4770590.html',
-        'https://vnexpress.net/10-cach-giam-roi-loan-tieu-hoa-4770010.html',
-        'https://vnexpress.net/polyp-dai-trang-4769281.html'
-        'https://vnexpress.net/thay-van-hai-la-cho-benh-nhan-thap-tim-lau-nam-4786873.html',
-        'https://vnexpress.net/nhan-giap-tien-trien-ung-thu-4785610.html',
-        'https://vnexpress.net/huyet-khoi-tinh-mach-dui-do-hoi-chung-may-thurner-4785159.html',
-        'https://vnexpress.net/cuong-giap-loi-mat-do-benh-basedow-4784710.html',
-        'https://vnexpress.net/phu-bach-huyet-khien-chan-trai-to-gap-doi-chan-phai-4782505.html',
-        'https://vnexpress.net/mo-bac-cau-mach-vanh-khi-trai-tim-van-dap-4781937.html',
-        'https://vnexpress.net/nguoi-dan-ong-nang-110-kg-bien-chung-nhoi-mau-co-tim-4781306.html',
-        'https://vnexpress.net/nong-mach-cuu-cu-ba-97-tuoi-nhoi-mau-co-tim-4781084.html',
-        'https://vnexpress.net/mang-nhen-tinh-mach-bao-vay-u-long-nguc-4780961.html',
-        'https://vnexpress.net/kham-viem-xoang-phat-hien-u-long-nguc-to-bang-qua-le-4779743.html',
-        'https://vnexpress.net/bit-lo-thong-tim-cho-nguoi-mac-nhieu-benh-nen-4779161.html',
-        'https://vnexpress.net/so-cuu-dung-cach-nguoi-dot-quy-nhoi-mau-co-tim-4779280.html'
-        'https://vnexpress.net/mo-tham-my-nhan-tuyen-giap-4776624.html',
-        'https://vnexpress.net/chan-sung-to-boi-huyet-khoi-tinh-mach-dui-4776390.html',
-        'https://vnexpress.net/chua-ung-thu-phoi-phat-hien-them-benh-mach-vanh-4775908.html',
-        'https://vnexpress.net/ba-nhanh-mach-vanh-tac-nghen-do-tang-cholesterol-mau-di-truyen-4774917.html',
-        'https://vnexpress.net/6-cach-giam-cholesterol-4774356.html',
-        'https://vnexpress.net/u-tim-4773956.html',
-        'https://vnexpress.net/u-o-dinh-phoi-to-gap-doi-sau-16-nam-4773348.html',
-        'https://vnexpress.net/suy-tim-sau-nhieu-nam-suy-than-man-4773335.html',
-        'https://vnexpress.net/huyet-ap-tay-trai-thap-hon-tay-phai-do-hep-mach-mau-nao-4770925.html',
-        'https://vnexpress.net/huyet-ap-cao-khong-ha-do-boc-tach-dong-mach-chu-4771203.html',
-        'https://vnexpress.net/song-khoe-sau-13-nam-suy-tim-4771677.html',
-        'https://vnexpress.net/cu-ong-ngung-tim-do-bien-chung-viem-co-tim-4772887.html',
-        'https://vnexpress.net/khi-nao-can-phau-thuat-ho-van-tim-4767272.html',
-        'https://vnexpress.net/phau-thuat-boc-xo-vua-dong-mach-ngan-dot-quy-4767563.html',
-        'https://vnexpress.net/tuong-ung-thu-hoa-nhan-giap-lanh-tinh-4767968.html',
-        'https://vnexpress.net/kho-tho-do-vo-ken-khi-phoi-4769681.html',
-        'https://vnexpress.net/dau-nguc-canh-bao-cac-benh-gi-4764693.html',
-        'https://vnexpress.net/45-phut-cuu-nguoi-dan-ong-nhoi-mau-co-tim-4764968.html',
-        'https://vnexpress.net/te-tay-boi-hoi-chung-loi-thoat-long-nguc-4766033.html',
-        'https://vnexpress.net/tac-nghen-dong-mach-khien-nguoi-dan-ong-suyt-nhoi-mau-co-tim-4766297.html',
-        'https://vnexpress.net/nhan-giap-bien-thanh-buou-to-sau-30-nam-4762262.html',
-        'https://vnexpress.net/nguc-lom-thanh-ho-2-cm-4762320.html',
-        'https://vnexpress.net/dau-nguc-trai-bao-hieu-nhoi-mau-co-tim-ma-khong-biet-4763777.html',
-        'https://vnexpress.net/ba-yeu-to-lam-tang-nguy-co-mac-benh-tim-o-nguoi-tre-4763953.html',
-        'https://vnexpress.net/thuoc-la-hai-tim-the-nao-4760035.html',
-        'https://vnexpress.net/bat-ngo-phat-hien-nhoi-mau-co-tim-tham-lang-4760662.html',
-        'https://vnexpress.net/quy-tac-5-an-5-tranh-giup-dieu-hoa-mo-mau-4761003.html',
-        'https://vnexpress.net/tai-sao-suy-tinh-mach-chi-duoi-gay-te-chan-4761789.html',
-        'https://vnexpress.net/bom-keo-chua-suy-tinh-mach-cho-nguoi-phu-nu-4757624.html',
-        'https://vnexpress.net/viem-co-tim-cap-4752427.html',
-        'https://vnexpress.net/huyet-khoi-lap-day-mach-vanh-sau-20-nam-hut-thuoc-la-4758708.html',
-        'https://vnexpress.net/tai-hep-mach-vanh-sau-dat-stent-4759311.html',
-        'https://vnexpress.net/cong-nghe-cao-phau-thuat-mach-mau-cuu-mang-nhieu-ca-benh-nang-4756026.html',
-        'https://vnexpress.net/benh-mach-vanh-thu-pham-gay-nhoi-mau-co-tim-4756327.html',
-        'https://vnexpress.net/phu-chan-voi-sau-nhieu-nam-dieu-tri-ung-thu-4756741.html',
-        'https://vnexpress.net/thieu-nu-lom-nguc-nguy-co-suy-tim-4757083.html',
-        'https://vnexpress.net/co-phinh-to-boi-nang-giap-4741323.html',
-        'https://vnexpress.net/xep-phoi-gay-7-xuong-suon-sau-cu-nga-cau-thang-4740863.html',
-        'https://vnexpress.net/an-uong-gi-de-tang-huyet-ap-4739520.html',
-        'https://vnexpress.net/dau-nguc-tuong-benh-tim-hoa-ung-thu-4739164.html',
-        'https://vnexpress.net/dau-hieu-nao-canh-bao-benh-tim-khong-nen-lo-la-4734788.html',
-        'https://vnexpress.net/bieu-hien-nao-o-da-mong-canh-bao-tim-khong-khoe-4735390.html',
-        'https://vnexpress.net/mach-mau-chan-tac-nghen-khien-cu-ong-khong-the-di-lai-4735573.html',
-        'https://vnexpress.net/hut-thuoc-la-co-lam-tang-nguy-co-dot-quy-4733800.html',
-        'https://vnexpress.net/u-giap-to-nhu-qua-trung-ga-gay-kho-nuot-4733274.html',
-        'https://vnexpress.net/dau-hieu-dien-hinh-cua-con-nhoi-mau-co-tim-4733879.html',
-        'https://vnexpress.net/tang-huyet-ap-co-dan-toi-dot-quy-4734412.html',
-        'https://vnexpress.net/mach-vanh-tac-nghen-nang-khien-cu-ong-suy-tim-4743670.html',
-        'https://vnexpress.net/mang-xo-vua-bit-kin-mach-vanh-nguoi-dan-ong-4744257.html',
-        'https://vnexpress.net/ai-nen-lam-xet-nghiem-dong-mau-4744603.html',
-        'https://vnexpress.net/thieu-mau-co-tim-nhieu-ngay-ma-khong-biet-4746074.html',
-        'https://vnexpress.net/co-phinh-to-boi-nang-giap-4741323.html',
-        'https://vnexpress.net/kha-nang-song-sot-cua-nguoi-bi-nhoi-mau-co-tim-4741899.html',
-        'https://vnexpress.net/nhip-tim-that-thuong-boi-mach-vanh-voi-hoa-nang-4742856.html',
-        'https://vnexpress.net/dat-stent-chu-y-khoi-thong-mach-mau-xoan-van-4743275.html',
-        'https://vnexpress.net/thoi-quen-ngu-tot-cho-tim-4743592.html',
-        'https://vnexpress.net/phat-hien-ung-thu-phoi-do-dot-bien-gene-4734643.html',
-        'https://vnexpress.net/so-cuu-nhoi-mau-co-tim-dung-cach-tranh-dot-tu-4738165.html',
-        'https://vnexpress.net/dong-mach-chu-nguc-phinh-to-doa-vo-4738101.html',
-        'https://vnexpress.net/suyt-dot-tu-do-tac-hai-mach-mau-lon-nuoi-tim-4737410.html',
-        'https://vnexpress.net/hai-lan-can-thiep-cho-nhac-cong-bi-nhoi-mau-co-tim-4737084.html',
-        'https://vnexpress.net/benh-tim-co-anh-huong-den-sinh-ly-4731612.html',
-        'https://vnexpress.net/nhieu-nguoi-bi-nhoi-mau-co-tim-ma-khong-biet-4732615.html',
-        'https://vnexpress.net/co-phinh-to-boi-nang-giap-4741323.html',
-        'https://vnexpress.net/mach-mau-chan-tac-nghen-khien-cu-ong-khong-the-di-lai-4735573.html',
-        'https://vnexpress.net/nhan-giap-bien-thanh-buou-to-sau-30-nam-4762262.html',
-        'https://vnexpress.net/nong-mach-cuu-cu-ba-97-tuoi-nhoi-mau-co-tim-4781084.html',
-        'https://vnexpress.net/mang-nhen-tinh-mach-bao-vay-u-long-nguc-4780961.html',
-        'https://vnexpress.net/quy-tac-5-an-5-tranh-giup-dieu-hoa-mo-mau-4761003.html',
-        'https://vnexpress.net/so-cuu-dung-cach-nguoi-dot-quy-nhoi-mau-co-tim-4779280.html'
-        'https://vnexpress.net/45-phut-cuu-nguoi-dan-ong-nhoi-mau-co-tim-4764968.html',
-        'https://vnexpress.net/thoi-quen-ngu-tot-cho-tim-4743592.html',
-        'https://vnexpress.net/huyet-ap-tay-trai-thap-hon-tay-phai-do-hep-mach-mau-nao-4770925.html',
-        'https://vnexpress.net/so-cuu-nhoi-mau-co-tim-dung-cach-tranh-dot-tu-4738165.html',
-        'https://vnexpress.net/ba-nhanh-mach-vanh-tac-nghen-do-tang-cholesterol-mau-di-truyen-4774917.html',
-        'https://vnexpress.net/te-tay-boi-hoi-chung-loi-thoat-long-nguc-4766033.html',
-        'https://vnexpress.net/thay-van-hai-la-cho-benh-nhan-thap-tim-lau-nam-4786873.html',
-        'https://vnexpress.net/huyet-khoi-tinh-mach-dui-do-hoi-chung-may-thurner-4785159.html',
-        'https://vnexpress.net/nhan-giap-tien-trien-ung-thu-4785610.html',
-        'https://vnexpress.net/mang-xo-vua-bit-kin-mach-vanh-nguoi-dan-ong-4744257.html',
-        'https://vnexpress.net/mo-bac-cau-mach-vanh-khi-trai-tim-van-dap-4781937.html',
-        'https://vnexpress.net/tai-hep-mach-vanh-sau-dat-stent-4759311.html',
-        'https://vnexpress.net/mo-tham-my-nhan-tuyen-giap-4776624.html',
-        'https://vnexpress.net/an-uong-gi-de-tang-huyet-ap-4739520.html',
-        'https://vnexpress.net/suyt-dot-tu-do-tac-hai-mach-mau-lon-nuoi-tim-4737410.html',
-        'https://vnexpress.net/u-o-dinh-phoi-to-gap-doi-sau-16-nam-4773348.html',
-        'https://vnexpress.net/cu-ong-ngung-tim-do-bien-chung-viem-co-tim-4772887.html',
-        'https://vnexpress.net/khi-nao-can-phau-thuat-ho-van-tim-4767272.html',
-        'https://vnexpress.net/phau-thuat-boc-xo-vua-dong-mach-ngan-dot-quy-4767563.html',
-        'https://vnexpress.net/thieu-mau-co-tim-nhieu-ngay-ma-khong-biet-4746074.html',
-        'https://vnexpress.net/chua-ung-thu-phoi-phat-hien-them-benh-mach-vanh-4775908.html',
-        'https://vnexpress.net/dau-nguc-canh-bao-cac-benh-gi-4764693.html',
-        'https://vnexpress.net/chan-sung-to-boi-huyet-khoi-tinh-mach-dui-4776390.html',
-        'https://vnexpress.net/mach-vanh-tac-nghen-nang-khien-cu-ong-suy-tim-4743670.html',
-        'https://vnexpress.net/tang-huyet-ap-co-dan-toi-dot-quy-4734412.html',
-        'https://vnexpress.net/bat-ngo-phat-hien-nhoi-mau-co-tim-tham-lang-4760662.html',
-        'https://vnexpress.net/phu-bach-huyet-khien-chan-trai-to-gap-doi-chan-phai-4782505.html',
-        'https://vnexpress.net/ba-yeu-to-lam-tang-nguy-co-mac-benh-tim-o-nguoi-tre-4763953.html',
-        'https://vnexpress.net/nhip-tim-that-thuong-boi-mach-vanh-voi-hoa-nang-4742856.html',
-        'https://vnexpress.net/bit-lo-thong-tim-cho-nguoi-mac-nhieu-benh-nen-4779161.html',
-        'https://vnexpress.net/cuong-giap-loi-mat-do-benh-basedow-4784710.html',
-        'https://vnexpress.net/kho-tho-do-vo-ken-khi-phoi-4769681.html',
-        'https://vnexpress.net/u-giap-to-nhu-qua-trung-ga-gay-kho-nuot-4733274.html',
-        'https://vnexpress.net/phu-chan-voi-sau-nhieu-nam-dieu-tri-ung-thu-4756741.html',
-        'https://vnexpress.net/dong-mach-chu-nguc-phinh-to-doa-vo-4738101.html',
-        'https://vnexpress.net/co-phinh-to-boi-nang-giap-4741323.html',
-        'https://vnexpress.net/thuoc-la-hai-tim-the-nao-4760035.html',
-        'https://vnexpress.net/phat-hien-ung-thu-phoi-do-dot-bien-gene-4734643.html',
-        'https://vnexpress.net/thieu-nu-lom-nguc-nguy-co-suy-tim-4757083.html',
-        'https://vnexpress.net/dat-stent-chu-y-khoi-thong-mach-mau-xoan-van-4743275.html',
-        'https://vnexpress.net/ai-nen-lam-xet-nghiem-dong-mau-4744603.html',
-        'https://vnexpress.net/hai-lan-can-thiep-cho-nhac-cong-bi-nhoi-mau-co-tim-4737084.html',
-        'https://vnexpress.net/benh-tim-co-anh-huong-den-sinh-ly-4731612.html',
-        'https://vnexpress.net/tac-nghen-dong-mach-khien-nguoi-dan-ong-suyt-nhoi-mau-co-tim-4766297.html',
-        'https://vnexpress.net/nhieu-nguoi-bi-nhoi-mau-co-tim-ma-khong-biet-4732615.html',
-        'https://vnexpress.net/xep-phoi-gay-7-xuong-suon-sau-cu-nga-cau-thang-4740863.html',
-        'https://vnexpress.net/6-cach-giam-cholesterol-4774356.html',
-        'https://vnexpress.net/bieu-hien-nao-o-da-mong-canh-bao-tim-khong-khoe-4735390.html',
-        'https://vnexpress.net/benh-mach-vanh-thu-pham-gay-nhoi-mau-co-tim-4756327.html',
-        'https://vnexpress.net/nguoi-dan-ong-nang-110-kg-bien-chung-nhoi-mau-co-tim-4781306.html',
-        'https://vnexpress.net/tai-sao-suy-tinh-mach-chi-duoi-gay-te-chan-4761789.html',
-        'https://vnexpress.net/viem-co-tim-cap-4752427.html',
-        'https://vnexpress.net/suy-tim-sau-nhieu-nam-suy-than-man-4773335.html',
-        'https://vnexpress.net/nguc-lom-thanh-ho-2-cm-4762320.html',
-        'https://vnexpress.net/dau-hieu-nao-canh-bao-benh-tim-khong-nen-lo-la-4734788.html',
-        'https://vnexpress.net/hut-thuoc-la-co-lam-tang-nguy-co-dot-quy-4733800.html',
-        'https://vnexpress.net/kham-viem-xoang-phat-hien-u-long-nguc-to-bang-qua-le-4779743.html',
-        'https://vnexpress.net/u-tim-4773956.html',
-        'https://vnexpress.net/huyet-khoi-lap-day-mach-vanh-sau-20-nam-hut-thuoc-la-4758708.html',
-        'https://vnexpress.net/cong-nghe-cao-phau-thuat-mach-mau-cuu-mang-nhieu-ca-benh-nang-4756026.html',
-        'https://vnexpress.net/huyet-ap-cao-khong-ha-do-boc-tach-dong-mach-chu-4771203.html',
-        'https://vnexpress.net/dau-nguc-trai-bao-hieu-nhoi-mau-co-tim-ma-khong-biet-4763777.html',
-        'https://vnexpress.net/dau-hieu-dien-hinh-cua-con-nhoi-mau-co-tim-4733879.html',
-        'https://vnexpress.net/dau-nguc-tuong-benh-tim-hoa-ung-thu-4739164.html',
-        'https://vnexpress.net/kha-nang-song-sot-cua-nguoi-bi-nhoi-mau-co-tim-4741899.html',
-        'https://vnexpress.net/tuong-ung-thu-hoa-nhan-giap-lanh-tinh-4767968.html',
-        'https://vnexpress.net/song-khoe-sau-13-nam-suy-tim-4771677.html',
-        'https://vnexpress.net/bom-keo-chua-suy-tinh-mach-cho-nguoi-phu-nu-4757624.html',
-"https://vnexpress.net/nhung-benh-tinh-duc-co-the-lay-qua-nu-hon-4783134.html",
-        "https://vnexpress.net/tai-sao-dung-bao-cao-su-van-co-the-lay-benh-tinh-duc-4786939.html",
-        "https://vnexpress.net/7-lam-tuong-ve-benh-tinh-duc-4777324.html",
-        "https://vnexpress.net/hpv-co-lay-khi-hon-4775535.html",
-        "https://vnexpress.net/mun-coc-sinh-duc-4766613.html",
-        "https://vnexpress.net/cac-benh-tinh-duc-gay-trieu-chung-giong-cum-4770334.html",
-        "https://vnexpress.net/ly-do-nen-xet-nghiem-benh-tinh-duc-4743891.html",
-        "https://vnexpress.net/5-trieu-chung-canh-bao-benh-tinh-duc-o-nam-gioi-4763239.html",
-        "https://vnexpress.net/tai-sao-khong-quan-he-ngoai-luong-van-nhiem-benh-tinh-duc-4752662.html",
-        "https://vnexpress.net/dau-hieu-benh-tinh-duc-o-nam-gioi-4747782.html",
-        "https://vnexpress.net/vo-chong-chung-thuy-sao-van-mac-giang-mai-4743086.html",
-        "https://vnexpress.net/ran-mu-4741780.html",
-        "https://vnexpress.net/nhung-benh-tinh-duc-co-the-lay-truyen-qua-da-4741528.html",
-        "https://vnexpress.net/benh-tinh-duc-nao-co-the-tai-nhiem-4736785.html",
-        "https://vnexpress.net/cac-benh-tinh-duc-co-the-truyen-tu-me-sang-con-4733275.html",
-        "https://vnexpress.net/tai-nhiem-giang-mai-khi-quan-he-dong-gioi-4732902.html",
-        "https://vnexpress.net/lam-gi-khi-nhiem-hpv-4730428.html",
-        "https://vnexpress.net/dau-hieu-canh-bao-giang-mai-4728592.html",
-        "https://vnexpress.net/hpv-duoc-chan-doan-nhu-the-nao-4722550.html",
-        "https://vnexpress.net/nhung-lam-tuong-thuong-gap-ve-hpv-4719413.html",
-        "https://vnexpress.net/bat-ngo-phat-hien-giang-mai-khi-san-con-tuoi-rong-4708297.html",
-        "https://vnexpress.net/nhung-benh-tinh-duc-co-the-gay-dau-hong-4702431.html",
-        "https://vnexpress.net/dau-hieu-mac-benh-lau-4700143.html",
-        "https://vnexpress.net/dau-hieu-nhan-biet-benh-tinh-duc-va-nhiem-trung-tiet-nieu-4654553.html",
-        "https://vnexpress.net/vi-sao-benh-giang-mai-kho-phat-hien-4659843.html",
-        "https://vnexpress.net/giang-mai-benh-tinh-duc-lam-giam-kha-nang-co-thai-4658528.html",
-        "https://vnexpress.net/hon-co-lay-benh-tinh-duc-4646388.html",
-        "https://vnexpress.net/mun-rop-sinh-duc-co-tai-phat-sau-dieu-tri-4642557.html",
-        "https://vnexpress.net/mun-mem-o-vung-kin-la-benh-gi-4635681.html",
-        "https://vnexpress.net/te-cung-chan-do-mac-giang-mai-4635529.html",
-        "https://vnexpress.net/di-tieu-rat-phat-hien-benh-lau-4632746.html",
-        "https://vnexpress.net/mat-ngu-vi-ran-mu-lam-to-vung-kin-4631945.html",
-        "https://vnexpress.net/dau-hieu-thuong-gap-cua-benh-lau-4623369.html",
-        "https://vnexpress.net/phan-biet-u-nhu-sinh-duc-va-sui-mao-ga-4621516.html",
-        "https://vnexpress.net/4-bien-chung-nguy-hiem-cua-giang-mai-4618518.html",
-        "https://vnexpress.net/tai-sao-khong-quan-he-van-co-the-mac-benh-tinh-duc-4605303.html",
-        "https://vnexpress.net/cac-benh-nhiem-trung-lay-qua-duong-tinh-duc-4608499.html",
-        "https://vnexpress.net/nhung-bien-chung-nguy-hiem-cua-benh-lau-4599514.html",
-        "https://vnexpress.net/can-benh-am-tham-gay-vo-sinh-o-ca-nam-va-nu-4599499.html",
-        "https://vnexpress.net/6-dieu-it-biet-ve-mun-rop-sinh-duc-4599449.html",
-        "https://vnexpress.net/sui-mao-ga-nguy-hiem-the-nao-4599443.html",
-'https://vnexpress.net/thai-nhi-16-tuan-mac-benh-tim-cao-doi-lot-cuu-4584122.html',
-        'https://vnexpress.net/be-14-thang-tuoi-xuat-vien-sau-3-ngay-mo-tim-ho-4580522.html',
-        'https://vnexpress.net/nhung-luu-y-khi-dua-tre-benh-tim-di-du-lich-4565020.html',
-        'https://vnexpress.net/co-nen-cho-tre-uong-thuoc-ngu-khi-sieu-am-tim-4522372.html',
-        'https://vnexpress.net/nguy-co-tang-huyet-ap-o-tre-dip-tet-nguyen-dan-4563164.html',
-        'https://vnexpress.net/co-nen-cho-tre-uong-thuoc-ngu-khi-sieu-am-tim-4522372.html',
-        'https://vnexpress.net/tre-5-tuoi-bi-thong-lien-that-co-the-phau-thuat-khong-4431093.html',
-        'https://vnexpress.net/cuu-song-be-6-ngay-tuoi-bi-teo-hep-mach-mau-chinh-nuoi-tim-4606819.html',
-        'https://vnexpress.net/be-trai-suy-tim-nang-do-khong-dieu-tri-benh-bam-sinh-4744799.html',
-        'https://vnexpress.net/cach-cham-soc-tre-mac-benh-tim-bam-sinh-4442021.html',
-        'https://vnexpress.net/7-thoi-quen-tot-cho-tim-4433939.html',
-        'https://vnexpress.net/be-hay-om-vat-chay-nhay-met-phat-hien-benh-tim-bam-sinh-4611925.html',
-        'https://vnexpress.net/cach-giup-kiem-soat-huyet-ap-o-tre-4431088.html',
-        'https://vnexpress.net/sieu-am-thai-co-phat-hien-benh-tim-bam-sinh-khong-4609642.html',
-        'https://vnexpress.net/be-hay-om-vat-chay-nhay-met-phat-hien-benh-tim-bam-sinh-4611925.html',
-        'https://vnexpress.net/cach-cham-soc-tre-mac-benh-tim-bam-sinh-4442021.html',
-        'https://vnexpress.net/nhung-luu-y-khi-dua-tre-benh-tim-di-du-lich-4565020.html',
-        'https://vnexpress.net/ca-mo-tim-be-gai-tu-ky-can-nao-bac-si-4691484.html',
-        'https://vnexpress.net/be-benh-tim-bam-sinh-hiem-gap-song-vui-khoe-khong-can-phau-thuat-4447742.html',
-        'https://vnexpress.net/be-trai-met-moi-kho-tho-di-kham-phat-hien-nguc-lom-sau-4553109.html',
-        'https://vnexpress.net/nhung-loai-trai-cay-tot-cho-tim-mach-cua-tre-4431180.html',
-        'https://vnexpress.net/ky-thuat-sieu-am-giup-phat-hien-di-tat-tim-bam-sinh-nguy-hiem-4452325.html',
-        'https://vnexpress.net/thoi-diem-nao-tam-soat-di-tat-tim-bam-sinh-chinh-xac-4547227.html',
-        'https://vnexpress.net/nhung-thoi-quen-xau-gay-hai-cho-trai-tim-cua-tre-4534733.html',
-        'https://vnexpress.net/o-nhiem-khong-khi-gay-hai-cho-tim-tre-nho-4431181.html',
-        'https://vnexpress.net/can-lam-gi-khi-tre-tang-huyet-ap-4541930.html',
-        'https://vnexpress.net/ky-thuat-sieu-am-giup-phat-hien-di-tat-tim-bam-sinh-nguy-hiem-4452325.html',
-        'https://vnexpress.net/be-hay-om-vat-chay-nhay-met-phat-hien-benh-tim-bam-sinh-4611925.html',
-        'https://vnexpress.net/cach-giup-kiem-soat-huyet-ap-o-tre-4431088.html',
-        'https://vnexpress.net/cach-cham-soc-tre-mac-benh-tim-bam-sinh-4442021.html',
-        'https://vnexpress.net/tre-beo-phi-de-bi-tang-huyet-ap-mo-mau-cao-4476742.html',
-        'https://vnexpress.net/be-benh-tim-bam-sinh-hiem-gap-song-vui-khoe-khong-can-phau-thuat-4447742.html',
-        'https://vnexpress.net/sieu-am-thai-co-phat-hien-benh-tim-bam-sinh-khong-4609642.html',
-        'https://vnexpress.net/be-gai-viem-co-tim-sau-hai-tuan-sot-sieu-vi-4758207.html',
-        'https://vnexpress.net/tre-sau-khi-mo-tim-co-choi-the-thao-duoc-khong-4604722.html',
-        'https://vnexpress.net/thai-nhi-mac-tu-chung-fallot-co-nguy-hiem-khong-4448702.html',
-        'https://vnexpress.net/nguy-co-tang-huyet-ap-o-tre-dip-tet-nguyen-dan-4563164.html',
-        'https://vnexpress.net/nhung-dau-hieu-tim-mach-bat-thuong-o-tre-sau-tiem-vaccine-covid-19-4455407.html',
-        'https://vnexpress.net/7-thoi-quen-tot-cho-tim-4433939.html',
-        'https://vnexpress.net/tam-soat-benh-tim-bam-sinh-tu-bao-thai-tang-co-hoi-song-cho-tre-4631623.html',
-        'https://vnexpress.net/tre-mac-benh-tim-can-luu-y-gi-khi-tiem-vaccine-covid-19-4455017.html',
-        'https://vnexpress.net/me-tieu-duong-thai-ky-co-sinh-con-di-tat-4678643.html',
-        'https://vnexpress.net/phat-hien-tim-bam-sinh-tu-bao-thai-giup-tre-chao-doi-an-toan-4513144.html',
-        'https://vnexpress.net/nguy-co-tang-huyet-ap-o-tre-dip-tet-nguyen-dan-4563164.html',
-        'https://vnexpress.net/o-nhiem-khong-khi-gay-hai-cho-tim-tre-nho-4431181.html',
-        'https://vnexpress.net/trieu-chung-nhan-biet-benh-tim-mach-o-tre-em-4511587.html',
-        'https://vnexpress.net/nguc-lom-bam-sinh-khi-nao-can-phau-thuat-4609009.html',
-        'https://vnexpress.net/nhung-dau-hieu-tim-mach-bat-thuong-o-tre-sau-tiem-vaccine-covid-19-4455407.html',
-        'https://vnexpress.net/thai-nhi-mac-tu-chung-fallot-co-nguy-hiem-khong-4448702.html',
-        'https://vnexpress.net/7-thoi-quen-tot-cho-tim-4433939.html',
-        'https://vnexpress.net/tre-beo-phi-de-bi-tang-huyet-ap-mo-mau-cao-4476742.html',
-        'https://vnexpress.net/nguy-co-tang-huyet-ap-o-tre-dip-tet-nguyen-dan-4563164.html',
-        'https://vnexpress.net/nhung-luu-y-khi-dua-tre-benh-tim-di-du-lich-4565020.html',
-        'https://vnexpress.net/nam-gioi-uong-ruou-co-the-gay-di-tat-tim-bam-sinh-o-con-4431221.html',
-        'https://vnexpress.net/nam-gioi-uong-ruou-co-the-gay-di-tat-tim-bam-sinh-o-con-4431221.html',
-        'https://vnexpress.net/be-trai-suy-tim-nang-do-khong-dieu-tri-benh-bam-sinh-4744799.html',
-        'https://vnexpress.net/cuu-song-be-6-ngay-tuoi-bi-teo-hep-mach-mau-chinh-nuoi-tim-4606819.html',
-        'https://vnexpress.net/tam-soat-tu-bao-thai-cuu-nhieu-tre-mac-benh-tim-bam-sinh-4438718.html',
-        'https://vnexpress.net/nguc-lom-bam-sinh-khi-nao-can-phau-thuat-4609009.html',
-        'https://vnexpress.net/tre-mac-benh-tim-can-luu-y-gi-khi-tiem-vaccine-covid-19-4455017.html',
-        'https://vnexpress.net/can-lam-gi-khi-tre-tang-huyet-ap-4541930.html',
-        'https://vnexpress.net/tam-soat-benh-tim-bam-sinh-tu-bao-thai-tang-co-hoi-song-cho-tre-4631623.html',
-        'https://vnexpress.net/benh-tim-bam-sinh-co-chua-duoc-khong-4608583.html',
-        'https://vnexpress.net/be-benh-tim-bam-sinh-hiem-gap-song-vui-khoe-khong-can-phau-thuat-4447742.html',
-        'https://vnexpress.net/be-trai-met-moi-kho-tho-di-kham-phat-hien-nguc-lom-sau-4553109.html',
-        'https://vnexpress.net/nhung-loai-trai-cay-tot-cho-tim-mach-cua-tre-4431180.html',
-        'https://vnexpress.net/benh-tim-bam-sinh-co-chua-duoc-khong-4608583.html',
-        'https://vnexpress.net/be-trai-met-moi-kho-tho-di-kham-phat-hien-nguc-lom-sau-4553109.html',
-        'https://vnexpress.net/be-hay-om-vat-chay-nhay-met-phat-hien-benh-tim-bam-sinh-4611925.html',
-        'https://vnexpress.net/kham-ho-hap-phat-hien-benh-tim-bam-sinh-4615410.html',
-        'https://vnexpress.net/phat-hien-tim-bam-sinh-tu-bao-thai-giup-tre-chao-doi-an-toan-4513144.html',
-        'https://vnexpress.net/phat-hien-tim-bam-sinh-tu-bao-thai-giup-tre-chao-doi-an-toan-4513144.html',
-        'https://vnexpress.net/nhung-luu-y-khi-dua-tre-benh-tim-di-du-lich-4565020.html',
-        'https://vnexpress.net/be-gai-viem-co-tim-sau-hai-tuan-sot-sieu-vi-4758207.html',
-        'https://vnexpress.net/co-nen-cho-tre-uong-thuoc-ngu-khi-sieu-am-tim-4522372.html',
-        'https://vnexpress.net/be-14-thang-tuoi-xuat-vien-sau-3-ngay-mo-tim-ho-4580522.html',
-        'https://vnexpress.net/can-lam-gi-khi-tre-tang-huyet-ap-4541930.html',
-        'https://vnexpress.net/sieu-am-phat-hien-som-di-tat-tim-bam-sinh-o-tre-4447575.html',
-        'https://vnexpress.net/cuu-song-be-6-ngay-tuoi-bi-teo-hep-mach-mau-chinh-nuoi-tim-4606819.html',
-        'https://vnexpress.net/nhung-dau-hieu-tim-mach-bat-thuong-o-tre-sau-tiem-vaccine-covid-19-4455407.html',
-        'https://vnexpress.net/ky-thuat-sieu-am-giup-phat-hien-di-tat-tim-bam-sinh-nguy-hiem-4452325.html',
-        'https://vnexpress.net/cuu-song-be-6-ngay-tuoi-bi-teo-hep-mach-mau-chinh-nuoi-tim-4606819.html',
-        'https://vnexpress.net/tre-5-tuoi-bi-thong-lien-that-co-the-phau-thuat-khong-4431093.html',
-        'https://vnexpress.net/be-14-thang-tuoi-xuat-vien-sau-3-ngay-mo-tim-ho-4580522.html',
-        'https://vnexpress.net/tre-mac-benh-tim-can-luu-y-gi-khi-tiem-vaccine-covid-19-4455017.html',
-        'https://vnexpress.net/ky-thuat-sieu-am-giup-phat-hien-di-tat-tim-bam-sinh-nguy-hiem-4452325.html',
-        'https://vnexpress.net/thai-nhi-16-tuan-mac-benh-tim-cao-doi-lot-cuu-4584122.html',
-        'https://vnexpress.net/me-tieu-duong-thai-ky-co-sinh-con-di-tat-4678643.html',
-        'https://vnexpress.net/tre-mac-benh-tim-can-luu-y-gi-khi-tiem-vaccine-covid-19-4455017.html',
-        'https://vnexpress.net/nhung-loai-trai-cay-tot-cho-tim-mach-cua-tre-4431180.html',
-        'https://vnexpress.net/benh-tim-bam-sinh-co-chua-duoc-khong-4608583.html',
-        'https://vnexpress.net/nhung-thoi-quen-xau-gay-hai-cho-trai-tim-cua-tre-4534733.html',
-        'https://vnexpress.net/7-thoi-quen-tot-cho-tim-4433939.html',
-        'https://vnexpress.net/kham-ho-hap-phat-hien-benh-tim-bam-sinh-4615410.html',
-        'https://vnexpress.net/can-thiep-ngay-sau-sinh-cho-be-gai-di-tat-tim-nang-4742387.html',
-        'https://vnexpress.net/tre-sau-khi-mo-tim-co-choi-the-thao-duoc-khong-4604722.html',
-        'https://vnexpress.net/cach-cham-soc-tre-mac-benh-tim-bam-sinh-4442021.html',
-        'https://vnexpress.net/ca-mo-tim-be-gai-tu-ky-can-nao-bac-si-4691484.html',
-        'https://vnexpress.net/me-tieu-duong-thai-ky-co-sinh-con-di-tat-4678643.html',
-        'https://vnexpress.net/thai-nhi-16-tuan-mac-benh-tim-cao-doi-lot-cuu-4584122.html',
-        'https://vnexpress.net/cach-giup-kiem-soat-huyet-ap-o-tre-4431088.html',
-        'https://vnexpress.net/can-thiep-ngay-sau-sinh-cho-be-gai-di-tat-tim-nang-4742387.html',
-        'https://vnexpress.net/7-thoi-quen-tot-cho-tim-4433939.html',
-        'https://vnexpress.net/tre-beo-phi-de-bi-tang-huyet-ap-mo-mau-cao-4476742.html',
-        'https://vnexpress.net/tam-soat-benh-tim-bam-sinh-tu-bao-thai-tang-co-hoi-song-cho-tre-4631623.html',
-        'https://vnexpress.net/tre-mac-benh-tim-can-luu-y-gi-khi-tiem-vaccine-covid-19-4455017.html',
-        'https://vnexpress.net/be-trai-suy-tim-nang-do-khong-dieu-tri-benh-bam-sinh-4744799.html',
-        'https://vnexpress.net/sieu-am-thai-co-phat-hien-benh-tim-bam-sinh-khong-4609642.html',
-        'https://vnexpress.net/dau-hieu-bien-phap-phong-ngua-benh-tim-bam-sinh-cho-tre-4253677.html',
-        'https://vnexpress.net/ca-mo-tim-be-gai-tu-ky-can-nao-bac-si-4691484.html',
-        'https://vnexpress.net/co-nen-cho-tre-uong-thuoc-ngu-khi-sieu-am-tim-4522372.html',
-        'https://vnexpress.net/sieu-am-phat-hien-som-di-tat-tim-bam-sinh-o-tre-4447575.html',
-        'https://vnexpress.net/can-thiep-ngay-sau-sinh-cho-be-gai-di-tat-tim-nang-4742387.html',
-        'https://vnexpress.net/tam-soat-benh-tim-bam-sinh-tu-bao-thai-tang-co-hoi-song-cho-tre-4631623.html',
-        'https://vnexpress.net/tre-beo-phi-de-bi-tang-huyet-ap-mo-mau-cao-4476742.html',
-        'https://vnexpress.net/thoi-diem-nao-tam-soat-di-tat-tim-bam-sinh-chinh-xac-4547227.html',
-        'https://vnexpress.net/thai-nhi-mac-tu-chung-fallot-co-nguy-hiem-khong-4448702.html',
-        'https://vnexpress.net/can-trong-benh-tang-huyet-ap-o-tre-4488907.html',
-        'https://vnexpress.net/nhung-dau-hieu-tim-mach-bat-thuong-o-tre-sau-tiem-vaccine-covid-19-4455407.html',
-        'https://vnexpress.net/dau-hieu-bien-phap-phong-ngua-benh-tim-bam-sinh-cho-tre-4253677.html',
-        'https://vnexpress.net/tre-beo-phi-de-bi-tang-huyet-ap-mo-mau-cao-4476742.html',
-        'https://vnexpress.net/ky-thuat-sieu-am-giup-phat-hien-di-tat-tim-bam-sinh-nguy-hiem-4452325.html',
-        'https://vnexpress.net/nam-gioi-uong-ruou-co-the-gay-di-tat-tim-bam-sinh-o-con-4431221.html',
-        'https://vnexpress.net/be-14-thang-tuoi-xuat-vien-sau-3-ngay-mo-tim-ho-4580522.html',
-        'https://vnexpress.net/tre-mac-benh-tim-can-luu-y-gi-khi-tiem-vaccine-covid-19-4455017.html',
-        'https://vnexpress.net/tam-soat-tu-bao-thai-cuu-nhieu-tre-mac-benh-tim-bam-sinh-4438718.html',
-        'https://vnexpress.net/tre-sau-khi-mo-tim-co-choi-the-thao-duoc-khong-4604722.html',
-        'https://vnexpress.net/phat-hien-tim-bam-sinh-tu-bao-thai-giup-tre-chao-doi-an-toan-4513144.html',
-        'https://vnexpress.net/benh-tim-bam-sinh-co-chua-duoc-khong-4608583.html',
-        'https://vnexpress.net/sieu-am-phat-hien-som-di-tat-tim-bam-sinh-o-tre-4447575.html',
-        'https://vnexpress.net/thoi-diem-nao-tam-soat-di-tat-tim-bam-sinh-chinh-xac-4547227.html',
-        'https://vnexpress.net/tre-sau-khi-mo-tim-co-choi-the-thao-duoc-khong-4604722.html',
-        'https://vnexpress.net/ca-mo-tim-be-gai-tu-ky-can-nao-bac-si-4691484.html',
-        'https://vnexpress.net/kham-ho-hap-phat-hien-benh-tim-bam-sinh-4615410.html',
-        'https://vnexpress.net/kham-ho-hap-phat-hien-benh-tim-bam-sinh-4615410.html',
-        'https://vnexpress.net/trieu-chung-nhan-biet-benh-tim-mach-o-tre-em-4511587.html',
-        'https://vnexpress.net/thai-nhi-mac-tu-chung-fallot-co-nguy-hiem-khong-4448702.html',
-        'https://vnexpress.net/dau-hieu-bien-phap-phong-ngua-benh-tim-bam-sinh-cho-tre-4253677.html',
-        'https://vnexpress.net/sieu-am-phat-hien-som-di-tat-tim-bam-sinh-o-tre-4447575.html',
-        'https://vnexpress.net/sieu-am-thai-co-phat-hien-benh-tim-bam-sinh-khong-4609642.html',
-        'https://vnexpress.net/tre-5-tuoi-bi-thong-lien-that-co-the-phau-thuat-khong-4431093.html',
-        'https://vnexpress.net/tre-5-tuoi-bi-thong-lien-that-co-the-phau-thuat-khong-4431093.html',
-        'https://vnexpress.net/can-lam-gi-khi-tre-tang-huyet-ap-4541930.html',
-        'https://vnexpress.net/tam-soat-tu-bao-thai-cuu-nhieu-tre-mac-benh-tim-bam-sinh-4438718.html',
-        'https://vnexpress.net/nhung-thoi-quen-xau-gay-hai-cho-trai-tim-cua-tre-4534733.html',
-        'https://vnexpress.net/o-nhiem-khong-khi-gay-hai-cho-tim-tre-nho-4431181.html',
-        'https://vnexpress.net/thoi-diem-nao-tam-soat-di-tat-tim-bam-sinh-chinh-xac-4547227.html',
-        'https://vnexpress.net/be-gai-viem-co-tim-sau-hai-tuan-sot-sieu-vi-4758207.html',
-        'https://vnexpress.net/can-trong-benh-tang-huyet-ap-o-tre-4488907.html',
-        'https://vnexpress.net/nhung-thoi-quen-xau-gay-hai-cho-trai-tim-cua-tre-4534733.html',
-        'https://vnexpress.net/cach-giup-kiem-soat-huyet-ap-o-tre-4431088.html',
-        'https://vnexpress.net/tre-mac-benh-tim-can-luu-y-gi-khi-tiem-vaccine-covid-19-4455017.html',
-        'https://vnexpress.net/be-trai-suy-tim-nang-do-khong-dieu-tri-benh-bam-sinh-4744799.html',
-        'https://vnexpress.net/o-nhiem-khong-khi-gay-hai-cho-tim-tre-nho-4431181.html',
-        'https://vnexpress.net/cuu-song-be-6-ngay-tuoi-bi-teo-hep-mach-mau-chinh-nuoi-tim-4606819.html',
-        'https://vnexpress.net/tre-mac-benh-tim-can-luu-y-gi-khi-tiem-vaccine-covid-19-4455017.html',
-        'https://vnexpress.net/thai-nhi-mac-tu-chung-fallot-co-nguy-hiem-khong-4448702.html',
-        'https://vnexpress.net/benh-tim-bam-sinh-co-chua-duoc-khong-4608583.html',
-        'https://vnexpress.net/benh-tim-bam-sinh-co-chua-duoc-khong-4608583.html',
-        'https://vnexpress.net/thai-nhi-16-tuan-mac-benh-tim-cao-doi-lot-cuu-4584122.html',
-        'https://vnexpress.net/me-tieu-duong-thai-ky-co-sinh-con-di-tat-4678643.html',
-        'https://vnexpress.net/trieu-chung-nhan-biet-benh-tim-mach-o-tre-em-4511587.html',
-        'https://vnexpress.net/o-nhiem-khong-khi-gay-hai-cho-tim-tre-nho-4431181.html',
-        'https://vnexpress.net/be-trai-met-moi-kho-tho-di-kham-phat-hien-nguc-lom-sau-4553109.html',
-        'https://vnexpress.net/nguc-lom-bam-sinh-khi-nao-can-phau-thuat-4609009.html',
-        'https://vnexpress.net/nguc-lom-bam-sinh-khi-nao-can-phau-thuat-4609009.html',
-        'https://vnexpress.net/nhung-loai-trai-cay-tot-cho-tim-mach-cua-tre-4431180.html',
-        'https://vnexpress.net/tam-soat-tu-bao-thai-cuu-nhieu-tre-mac-benh-tim-bam-sinh-4438718.html',
-        'https://vnexpress.net/be-gai-viem-co-tim-sau-hai-tuan-sot-sieu-vi-4758207.html',
-        'https://vnexpress.net/nam-gioi-uong-ruou-co-the-gay-di-tat-tim-bam-sinh-o-con-4431221.html',
-        'https://vnexpress.net/nguy-co-tang-huyet-ap-o-tre-dip-tet-nguyen-dan-4563164.html',
-        'https://vnexpress.net/can-thiep-ngay-sau-sinh-cho-be-gai-di-tat-tim-nang-4742387.html',
-        'https://vnexpress.net/can-trong-benh-tang-huyet-ap-o-tre-4488907.html',
-        'https://vnexpress.net/tre-mac-benh-tim-can-luu-y-gi-khi-tiem-vaccine-covid-19-4455017.html',
-        'https://vnexpress.net/nhung-luu-y-khi-dua-tre-benh-tim-di-du-lich-4565020.html',
-        'https://vnexpress.net/trieu-chung-nhan-biet-benh-tim-mach-o-tre-em-4511587.html',
-        'https://vnexpress.net/dau-hieu-bien-phap-phong-ngua-benh-tim-bam-sinh-cho-tre-4253677.html',
-        'https://vnexpress.net/be-benh-tim-bam-sinh-hiem-gap-song-vui-khoe-khong-can-phau-thuat-4447742.html',
-        'https://vnexpress.net/can-trong-benh-tang-huyet-ap-o-tre-4488907.html',
-        'https://vnexpress.net/can-thiep-ngay-sau-sinh-cho-be-gai-di-tat-tim-nang-4742387.html',
-'https://vnexpress.net/nguyen-nhan-dau-bung-sau-an-trai-cay-4784293.html',
-        'https://vnexpress.net/suy-kiet-do-viem-loet-dai-trang-boi-nhiem-virus-4784665.html',
-        'https://vnexpress.net/nhung-mon-an-uong-co-the-gay-hoi-chung-ruot-kich-thich-4784879.html',
-        'https://vnexpress.net/khoi-nguy-co-ghep-gan-nho-thay-24-lit-huyet-tuong-4779669.html',
-        'https://vnexpress.net/cac-trieu-chung-tieu-hoa-canh-bao-benh-4785101.html',
-        'https://vnexpress.net/7-loai-tra-giup-giam-day-hoi-4785622.html',
-        'https://vnexpress.net/thu-pham-gay-ung-thu-truc-trang-4756695.html',
-        'https://vnexpress.net/an-chay-co-giam-nguy-co-ung-thu-duong-tieu-hoa-4757225.html',
-        'https://vnexpress.net/nhung-dieu-khong-nen-lam-sau-khi-noi-soi-tieu-hoa-4623953.html',
-        'https://vnexpress.net/noi-soi-dat-stent-cho-cu-ong-co-hai-khoi-u-ac-tinh-4624282.html',
-        'https://vnexpress.net/noi-soi-dat-stent-cho-cu-ong-co-hai-khoi-u-ac-tinh-4624282.html',
-        'https://vnexpress.net/tai-tao-thuc-quan-bang-ong-da-day-cho-nguoi-ung-thu-4628319.html'
-'https://vnexpress.net/ba-nhom-dinh-duong-trong-nuoc-ham-xuong-4787674.html',
-                  'https://vnexpress.net/thoai-hoa-khop-hang-4787803.html',
-                  'https://vnexpress.net/mac-viem-khop-co-duoc-an-trung-4786718.html',
-                  'https://vnexpress.net/7-bai-tap-hang-ngay-giam-dau-khop-4786066.html',
-                  'https://vnexpress.net/xet-nghiem-chan-doan-gout-the-nao-4786274.html',
-                  'https://vnexpress.net/nhung-loai-rau-cu-giau-canxi-giup-xuong-chac-khoe-4785249.html',
-                  'https://vnexpress.net/lieu-phap-nong-lanh-giup-giam-viem-khop-4784304.html',
-                  'https://vnexpress.net/tinh-the-gout-hinh-thanh-the-nao-4784069.html',
-                  'https://vnexpress.net/dong-cung-vai-canh-bao-viem-khop-4783639.html',
-                  'https://vnexpress.net/nhung-nguoi-nen-tiem-chat-nhon-cho-khop-goi-4782536.html',
-                  'https://vnexpress.net/bi-benh-gout-an-dua-duoc-khong-4780912.html',
-                  'https://vnexpress.net/5-bai-tap-nen-luu-y-khi-thoai-hoa-khop-goi-4779893.html',
-                  'https://vnexpress.net/thua-can-beo-phi-de-nang-xuong-khop-4779697.html',
-                  'https://vnexpress.net/stress-lam-tang-nguy-co-mac-benh-gout-the-nao-4776537.html',
-                  'https://vnexpress.net/thoai-hoa-khop-goi-co-chua-khoi-khong-4778170.html',
-                  'https://vnexpress.net/nhung-thuc-pham-de-gay-yeu-xuong-khop-4777939.html',
-                  'https://vnexpress.net/dau-long-ban-chan-khi-ngu-day-la-benh-gi-4776688.html',
-                  'https://vnexpress.net/cach-giam-kho-cung-khop-4775689.html',
-                  'https://vnexpress.net/tai-sao-phu-nu-de-mac-viem-khop-dang-thap-4775075.html',
-                  'https://vnexpress.net/4-giai-doan-cua-benh-gout-4774233.html',
-                  'https://vnexpress.net/thoai-hoa-khop-khac-viem-khop-dang-thap-the-nao-4773933.html',
-                  'https://vnexpress.net/kho-gap-duoi-ngon-tay-canh-bao-viem-bao-gan-4773690.html',
-                  'https://vnexpress.net/lam-dung-ruou-bia-huy-hoai-xuong-dui-4773008.html',
-                  'https://vnexpress.net/benh-gout-co-di-truyen-khong-4772016.html',
-                  'https://vnexpress.net/dau-hieu-thoai-hoa-khop-goi-theo-tung-giai-doan-4771268.html',
-                  'https://vnexpress.net/meo-ngu-ngon-cho-nguoi-benh-viem-khop-dang-thap-4770062.html',
-                  'https://vnexpress.net/6-bai-tap-tot-cho-nguoi-thoai-hoa-khop-goi-4769192.html',
-                  'https://vnexpress.net/5-hoat-dong-nguoi-benh-thoai-hoa-khop-goi-nen-tranh-4767998.html',
-                  'https://vnexpress.net/suyt-tan-phe-do-lo-la-chua-gout-4767567.html',
-                  'https://vnexpress.net/dau-bap-co-tot-cho-xuong-khop-4766795.html',
-                  'https://vnexpress.net/12-benh-co-dau-hieu-te-bi-tay-chan-4764764.html',
-                  'https://vnexpress.net/khi-nao-nen-tiem-chat-nhon-cho-khop-goi-4763264.html',
-                  'https://vnexpress.net/nguyen-nhan-nao-gay-tang-axit-uric-4761786.html',
-                  'https://vnexpress.net/nguyen-nhan-thuong-gap-gay-thoai-hoa-khop-4760491.html',
-                  'https://vnexpress.net/uong-ruou-bia-gay-hai-xuong-the-nao-4761080.html',
-                  'https://vnexpress.net/e-am-dau-nhuc-xuong-khop-canh-bao-benh-gi-4758494.html',
-                  'https://vnexpress.net/dau-nhuc-khop-hang-canh-bao-hoai-tu-chom-xuong-dui-4757294.html',
-                  'https://vnexpress.net/4-giai-doan-thoai-hoa-khop-4756072.html',
-                  'https://vnexpress.net/hai-chan-veo-45-do-do-thoai-hoa-khop-goi-4755700.html',
-                  'https://vnexpress.net/nguoi-benh-thoat-vi-dia-dem-nen-an-gi-4755001.html',
-                  'https://vnexpress.net/phan-biet-benh-gout-va-gia-gout-4751115.html',
-                  'https://vnexpress.net/lam-tuong-ve-thuc-pham-gay-viem-khop-4749578.html',
-                  'https://vnexpress.net/nhung-phuong-phap-hien-dai-dieu-tri-thoai-hoa-khop-goi-4749211.html',
-                  'https://vnexpress.net/het-dau-sau-10-nam-mac-benh-khop-hang-4748198.html',
-                  'https://vnexpress.net/ung-dung-tri-tue-nhan-tao-chan-doan-loang-xuong-4747702.html',
-                  'https://vnexpress.net/cach-chon-giay-cho-nguoi-viem-khop-ban-chan-4746480.html',
-                  'https://vnexpress.net/chay-bo-co-gay-thoai-hoa-khop-goi-4746675.html',
-                  'https://vnexpress.net/viem-khop-co-the-gay-tan-tat-khong-4745345.html',
-                  'https://vnexpress.net/khop-nao-de-bi-thoai-hoa-4743645.html',
-                  'https://vnexpress.net/nguy-co-ton-thuong-gan-co-tay-do-nghien-smartphone-4743939.html',
-'https://vnexpress.net/cu-ba-4-lan-thay-khop-hang-4784727.html',
-                  'https://vnexpress.net/khau-vet-rach-gan-co-vai-nguoi-dan-ong-4782050.html',
-                  'https://vnexpress.net/4-yeu-to-tang-do-ben-khop-nhan-tao-sau-phau-thuat-4780443.html',
-                  'https://vnexpress.net/ca-phau-thuat-gian-nan-thay-khop-cho-benh-nhan-mau-hiem-4777231.html',
-                  'https://vnexpress.net/day-chang-nhan-tao-giup-nguoi-benh-di-lai-sau-chan-thuong-4774692.html',
-                  'https://vnexpress.net/dau-co-chan-bat-ngo-phat-hien-u-xuong-4773295.html',
-                  'https://vnexpress.net/phuc-hoi-dang-di-sau-50-nam-gay-xuong-dui-4772173.html',
-                  'https://vnexpress.net/di-tat-thua-dinh-ngon-tay-o-tre-4770324.html',
-                  'https://vnexpress.net/dieu-tri-chan-chu-x-nhu-the-nao-4769571.html',
-                  'https://vnexpress.net/lam-the-nao-nhanh-phuc-hoi-sau-mo-thay-khop-hang-4767763.html',
-                  'https://vnexpress.net/thay-khop-goi-cho-nguoi-phu-nu-beo-phi-4766466.html',
-                  'https://vnexpress.net/mac-benh-truyen-nhiem-co-thay-khop-duoc-khong-4765970.html',
-                  'https://vnexpress.net/cu-ba-bien-chung-thuyen-tac-phoi-sau-thay-khop-hang-4765674.html',
-                  'https://vnexpress.net/nguyen-nhan-tai-dut-day-chang-sau-mo-4765139.html',
-                  'https://vnexpress.net/bi-lat-co-chan-thuong-xuyen-co-nguy-hiem-4763822.html',
-                  'https://vnexpress.net/5-di-tat-co-xuong-khop-bam-sinh-o-tre-nho-4762874.html',
-                  'https://vnexpress.net/manh-xuong-vo-ket-trong-khop-sau-cu-nga-xe-dap-4762265.html',
-                  'https://vnexpress.net/chuyen-vat-da-va-vet-loet-lo-xuong-4760697.html',
-                  'https://vnexpress.net/hai-chan-cong-veo-nang-do-khop-goi-thoai-hoa-4760111.html#vn_source=Folder-SucKhoe_CacBenh_BenhCoXuongKhop_ChanThuongChinhHinh&vn_campaign=Stream&vn_medium=Item-19&vn_term=Desktop&vn_thumb=1&vn_aid=1005274',
-                  'https://vnexpress.net/dut-day-chang-dau-goi-khi-nao-can-mo-4759275.html',
-                  'https://vnexpress.net/ung-thu-than-di-can-xuong-sau-10-thang-khong-dieu-tri-4757882.html',
-                  'https://vnexpress.net/thay-hai-khop-goi-sau-15-nam-thoai-hoa-4756823.html',
-                  'https://vnexpress.net/ba-lan-dut-day-chang-gay-teo-co-long-khop-goi-4755094.html',
-                  'https://vnexpress.net/cac-phuong-phap-chua-chan-vong-kieng-4753132.html',
-                  'https://vnexpress.net/rach-gan-vai-benh-thuong-gap-o-nguoi-cao-tuoi-4752362.html',
-                  'https://vnexpress.net/dau-co-chan-khi-chay-bo-4751932.html',
-                  'https://vnexpress.net/dut-ba-day-chang-khop-goi-sau-cu-nga-xe-4751420.html',
-                  'https://vnexpress.net/suyt-mat-ngon-tay-do-ket-vao-xe-day-hanh-ly-4750362.html',
-                  'https://vnexpress.net/thay-ca-hai-khop-goi-do-thoai-hoa-giai-doan-cuoi-4746224.html',
-                  'https://vnexpress.net/cach-giam-dau-sau-tai-tao-day-chang-4745891.html',
-                  'https://vnexpress.net/thay-khop-hang-co-tap-gym-duoc-khong-4744883.html',
-                  'https://vnexpress.net/thay-khop-bang-su-cho-nguoi-hoai-tu-chom-xuong-dui-4744524.html',
-                  'https://vnexpress.net/gay-xuong-don-khi-choi-bong-da-4743476.html',
-                  'https://vnexpress.net/thay-khop-hang-giai-phap-thay-the-chom-xuong-dui-hoai-tu-4742603.html',
-                  'https://vnexpress.net/thay-hai-khop-goi-trong-mot-thang-do-thoai-hoa-nang-4741007.html',
-                  'https://vnexpress.net/nghien-ruou-pha-huy-chom-xuong-dui-cua-nam-thanh-nien-4739409.html',
-                  'https://vnexpress.net/di-lai-binh-thuong-sau-40-nam-khap-khieng-nho-thay-khop-hang-4738543.html',
-                  'https://vnexpress.net/nhieu-nguoi-tre-phai-thay-khop-hang-4736469.html',
-                  'https://vnexpress.net/tri-hoan-dieu-tri-nam-thanh-nien-dut-lien-tiep-hai-day-chang-4735281.html',
-                  'https://vnexpress.net/bi-thoai-hoa-goi-co-nen-thay-khop-nhan-tao-4734933.html',
-                  'https://vnexpress.net/dau-ba-vai-canh-bao-nhieu-benh-4734455.html',
-                  'https://vnexpress.net/bien-dang-xuong-sau-10-nam-thoai-hoa-khop-hang-4733618.html',
-                  'https://vnexpress.net/bi-dien-giat-nga-gay-nat-khop-vai-4732182.html',
-                  'https://vnexpress.net/ca-si-quang-tu-bi-hoai-tu-chom-xuong-dui-4731274.html',
-                  'https://vnexpress.net/ba-lan-dut-day-chang-do-choi-the-thao-4730979.html',
-                  'https://vnexpress.net/tai-sao-van-dau-nhuc-sau-thay-khop-goi-4730562.html',
-                  'https://vnexpress.net/rach-lon-gan-vai-gay-teo-co-4730117.html',
-                  'https://vnexpress.net/thu-pham-gay-dau-bap-chan-4729099.html',
-                  'https://vnexpress.net/noi-xuong-dui-cho-hai-cu-ba-u90-4727676.html',
-                  'https://vnexpress.net/tai-sao-phai-tiem-chat-nhon-sau-phau-thuat-day-chang-4726668.html',
-                  'https://vnexpress.net/hoai-tu-chom-xuong-dui-hai-lan-thay-khop-hang-4726307.html',
-                  'https://vnexpress.net/di-lai-binh-thuong-sau-6-nam-thoai-hoa-khop-goi-4724450.html',
-                  'https://vnexpress.net/ly-do-phu-nu-mac-benh-xuong-khop-cao-hon-nam-gioi-4723810.html',
-'https://vnexpress.net/10-nam-thoat-vi-dia-dem-4787272.html',
-                  'https://vnexpress.net/ton-thuong-cot-song-do-nhiem-vi-khuan-an-thit-nguoi-4785175.html#vn_source=Folder-SucKhoe_CacBenh_BenhCoXuongKhop_ThanKinhCotSong&vn_campaign=Stream&vn_medium=Item-2&vn_term=Desktop&vn_thumb=1&vn_aid=1005275',
-                  'https://vnexpress.net/di-nhu-robot-do-viem-dinh-cot-song-4781479.html',
-                  'https://vnexpress.net/bi-thoat-vi-dia-dem-co-phai-uong-thuoc-suot-doi-khong-4781053.html',
-                  'https://vnexpress.net/cac-phuong-phap-dieu-tri-xep-dia-dem-4776119.html',
-                  'https://vnexpress.net/12-benh-co-dau-hieu-te-bi-tay-chan-4764764.html',
-                  'https://vnexpress.net/dau-hieu-canh-bao-viem-cot-song-dinh-khop-4759683.html',
-                  'https://vnexpress.net/6-bai-tap-giam-dau-than-kinh-toa-4758258.html',
-                  'https://vnexpress.net/meo-ngu-ngon-cho-nguoi-thoat-vi-dia-dem-4756466.html',
-                  'https://vnexpress.net/5-bai-tap-cai-thien-cong-veo-cot-song-4755692.html',
-                  'https://vnexpress.net/7-hoat-dong-thuong-ngay-nen-tranh-khi-thoat-vi-dia-dem-4754573.html',
-                  'https://vnexpress.net/xep-dot-song-chua-duoc-khong-4754034.html',
-                  'https://vnexpress.net/cac-bai-tap-giam-dau-cho-nguoi-thoat-vi-dia-dem-4752752.html',
-                  'https://vnexpress.net/thoat-vi-dia-dem-co-chay-bo-duoc-khong-4748728.html',
-                  'https://vnexpress.net/tu-the-nam-giam-dau-than-kinh-toa-4741969.html',
-                  'https://vnexpress.net/ai-de-bi-gai-cot-song-4740165.html',
-                  'https://vnexpress.net/loi-ich-khi-mo-noi-soi-thoat-vi-dia-dem-4724032.html',
-                  'https://vnexpress.net/nguyen-nhan-gay-truot-dot-song-4721785.html',
-                  'https://vnexpress.net/gu-gap-lung-hai-chi-em-khong-the-nam-ngua-4721369.html',
-                  'https://vnexpress.net/dau-dot-song-co-va-that-lung-la-benh-gi-4717325.html',
-                  'https://vnexpress.net/thoat-vi-dia-dem-chen-ep-than-kinh-4716821.html',
-                  'https://vnexpress.net/dau-hieu-u-cot-song-4716589.html',
-                  'https://vnexpress.net/cu-ong-101-tuoi-nga-xep-dot-song-4713736.html',
-                  'https://vnexpress.net/tri-hoan-mo-thoat-vi-dia-dem-nguoi-dan-ong-suyt-liet-4713313.html',
-                  'https://vnexpress.net/thoi-quen-ngay-tet-gay-dau-lung-4710200.html',
-                  'https://vnexpress.net/mo-thoat-vi-dia-dem-cho-nguoi-dai-thao-duong-4709246.html',
-                  'https://vnexpress.net/nguy-co-thoat-vi-dia-dem-khi-don-nha-don-tet-4708916.html',
-                  'https://vnexpress.net/10-nam-dau-vai-gay-phai-thay-dot-song-nhan-tao-4707665.html',
-                  'https://vnexpress.net/cu-ong-94-tuoi-xep-dot-song-do-vac-nang-4702742.html',
-                  'https://vnexpress.net/thoat-vi-dia-dem-co-do-thuong-cui-dau-4700957.html',
-                  'https://vnexpress.net/cac-dang-cong-veo-cot-song-thuong-gap-4696087.html',
-                  'https://vnexpress.net/cot-song-cau-tao-nhu-the-nao-4694465.html',
-                  'https://vnexpress.net/dau-hieu-canh-bao-thoat-vi-dia-dem-lung-4690213.html',
-                  'https://vnexpress.net/ngu-guc-tren-ban-hai-xuong-khop-4688322.html',
-                  'https://vnexpress.net/7-duong-chat-tot-cho-nguoi-benh-thoat-vi-dia-dem-4687454.html',
-                  'https://vnexpress.net/benh-cot-song-de-bi-chan-doan-nham-4683981.html',
-                  'https://vnexpress.net/khi-nao-can-phau-thuat-hep-ong-song-4683528.html',
-                  'https://vnexpress.net/chan-thuong-cot-song-anh-huong-the-nao-den-doi-song-chan-goi-4680852.html',
-                  'https://vnexpress.net/ba-cach-giup-dan-van-phong-tranh-dau-cot-song-4680066.html',
-                  'https://vnexpress.net/dau-hieu-liet-day-than-kinh-so-7-4677619.html',
-                  'https://vnexpress.net/dau-hieu-truot-dot-song-4676716.html',
-                  'https://vnexpress.net/dau-hong-khi-ngoi-la-trieu-chung-cua-benh-gi-4676387.html',
-                  'https://vnexpress.net/nguyen-nhan-gay-cung-co-4675245.html',
-                  'https://vnexpress.net/4-mon-the-thao-tot-cho-nguoi-dau-than-kinh-toa-4674769.html#vn_source=Folder-SucKhoe_CacBenh_BenhCoXuongKhop_ThanKinhCotSong&vn_campaign=Stream&vn_medium=Item-44&vn_term=Desktop&vn_thumb=1&vn_aid=1005275',
-                  'https://vnexpress.net/benh-tu-trieu-chung-dau-rat-lung-4667342.html',
-                  'https://vnexpress.net/chan-nang-nhu-deo-da-do-day-chang-cot-song-voi-hoa-4664024.html',
-                  'https://vnexpress.net/thoi-quen-khi-ngu-khong-tot-cho-cot-song-4662046.html',
-                  'https://vnexpress.net/chua-gu-lung-bang-vat-ly-tri-lieu-4660815.html',
-                  'https://vnexpress.net/thuong-xuyen-dau-lung-la-benh-gi-4660222.html',
-                  'https://vnexpress.net/bien-chung-nguy-hiem-cua-thoat-vi-dia-dem-4657149.html',
-                  'https://vnexpress.net/ngoi-cong-lung-hai-xuong-khop-4656733.html',
-                  'https://vnexpress.net/cac-benh-dau-co-tay-do-go-may-tinh-nhieu-4654524.html',
-                  'https://vnexpress.net/nhung-tu-the-ngoi-hoc-gay-cong-veo-cot-song-4651820.html',
-                  'https://vnexpress.net/dau-hieu-ton-thuong-cot-song-4650902.html',
-                  'https://vnexpress.net/5-loai-chan-thuong-co-thuong-gap-4647883.html'
-'https://vnexpress.net/ba-lam-tuong-thuong-gap-ve-phuc-hoi-chuc-nang-4778732.html',
-                  'https://vnexpress.net/sau-an-bao-lau-co-the-chay-bo-4774852.html',
-                  'https://vnexpress.net/meo-di-bo-giup-dot-chay-nhieu-calo-4763908.html',
-                  'https://vnexpress.net/an-gi-truoc-va-sau-chay-bo-4761547.html',
-                  'https://vnexpress.net/tai-sao-dau-that-lung-khi-chay-bo-4753611.html',
-                  'https://vnexpress.net/nguy-co-dut-day-chang-o-nguoi-choi-the-thao-nghiep-du-4750167.html',
-                  'https://vnexpress.net/5-meo-chay-bo-giam-can-nhanh-4747769.html',
-                  'https://vnexpress.net/chay-bo-the-nao-de-giam-can-4744527.html',
-                  'https://vnexpress.net/5-sai-lam-khi-chay-bo-gay-hai-xuong-khop-4737056.html',
-                  'https://vnexpress.net/5-bai-tap-tang-suc-manh-co-bap-tay-4732800.html',
-                  'https://vnexpress.net/chay-marathon-the-nao-khong-hai-xuong-khop-4732584.html',
-                  'https://vnexpress.net/6-thoi-quen-giup-nguoi-trung-nien-khoe-xuong-khop-4731582.html',
-                  'https://vnexpress.net/sau-mo-day-chang-bao-lau-co-the-choi-the-thao-4714520.html',
-                  'https://vnexpress.net/tap-luyen-tro-lai-sau-tet-can-luu-y-gi-4712967.html',
-                  'https://vnexpress.net/tai-sao-de-chan-thuong-khi-chay-bo-4704629.html',
-                  'https://vnexpress.net/loi-ich-cua-nang-ta-voi-xuong-khop-4703252.html',
-                  'https://vnexpress.net/ba-vi-tri-tren-co-the-de-chan-thuong-khi-tap-gym-4697638.html',
-                  'https://vnexpress.net/loi-ich-voi-xuong-khop-khi-di-bo-nhanh-4697172.html',
-                  'https://vnexpress.net/tap-hiit-giam-can-tang-co-4696353.html',
-                  'https://vnexpress.net/nguy-co-thoai-hoa-khop-do-chan-thuong-the-thao-4694080.html',
-                  'https://vnexpress.net/dau-hieu-canh-bao-ban-chay-bo-qua-suc-4693094.html',
-                  'https://vnexpress.net/bai-tap-giam-can-tang-co-khong-can-den-phong-gym-4690643.html',
-                  'https://vnexpress.net/lam-the-nao-chay-ben-khong-met-4689523.html',
-                  'https://vnexpress.net/dau-hieu-mat-co-bap-4689282.html',
-                  'https://vnexpress.net/dap-xe-tac-dong-den-co-the-the-nao-4685396.html',
-                  'https://vnexpress.net/loi-ich-khi-tap-kick-boxing-4683950.html',
-                  'https://vnexpress.net/loi-ich-khi-di-bo-buoi-sang-4682173.html',
-                  'https://vnexpress.net/luu-y-cho-nguoi-lan-dau-chay-marathon-4681821.html',
-                  'https://vnexpress.net/nhung-loi-tap-ta-de-mac-4680325.html',
-                  'https://vnexpress.net/loi-ich-khi-nang-ta-nhe-4679155.html',
-                  'https://vnexpress.net/co-nen-tap-ta-hang-ngay-de-tang-co-bap-4678393.html',
-                  'https://vnexpress.net/co-nen-chay-bo-khi-troi-lanh-4676885.html',
-                  'https://vnexpress.net/tap-pilates-tot-cho-xuong-khop-the-nao-4672222.html',
-                  'https://vnexpress.net/loi-ich-khi-tap-luyen-voi-doi-chan-tran-4671813.html',
-                  'https://vnexpress.net/loi-ich-khi-tap-plank-4671249.html',
-                  'https://vnexpress.net/loi-ich-khi-di-bo-lui-4670451.html',
-                  'https://vnexpress.net/ly-do-dau-xoc-hong-khi-tap-the-duc-4669894.html',
-                  'https://vnexpress.net/tai-sao-tap-the-duc-nhung-khong-do-mo-hoi-4663528.html',
-                  'https://vnexpress.net/tai-sao-long-ban-chan-dau-khi-ban-chay-bo-tren-may-4662841.html',
-                  'https://vnexpress.net/5-dong-tac-yoga-giup-ban-cang-gia-cang-deo-4660631.html',
-                  'https://vnexpress.net/tai-sao-ban-nen-khoi-dong-truoc-khi-tap-luyen-4659852.html',
-                  'https://vnexpress.net/ly-do-ban-nen-tap-the-duc-buoi-toi-4658203.html',
-                  'https://vnexpress.net/ban-da-tap-squat-dung-cach-4654057.html',
-                  'https://vnexpress.net/vi-sao-ban-thuong-nong-chan-khi-chay-bo-4652607.html',
-                  'https://vnexpress.net/an-gi-giam-dau-co-sau-khi-tap-the-duc-4651423.html',
-                  'https://vnexpress.net/chan-thuong-thuong-gap-khi-chay-bo-4651192.html',
-                  'https://vnexpress.net/tap-hit-dat-bao-nhieu-lan-mot-ngay-4647731.html',
-                  'https://vnexpress.net/sai-lam-khi-di-bo-tang-nguy-co-chan-thuong-4646574.html',
-                  'https://vnexpress.net/dap-xe-co-lam-to-bap-chan-4643568.html',
-                  'https://vnexpress.net/chay-duoi-nuoc-mon-the-thao-co-loi-cho-xuong-khop-4643327.html',
-                  'https://vnexpress.net/trac-nghiem-chon-mon-the-thao-tot-cho-xuong-khop-4634156.html',
-                  'https://vnexpress.net/vi-sao-dau-dau-sau-khi-tap-the-duc-4632139.html',
-                  'https://vnexpress.net/cach-chon-giay-chay-bo-phu-hop-4629398.html',
-                  'https://vnexpress.net/cac-bai-tap-yoga-giup-lung-thang-4626339.html',
-'https://vnexpress.net/mac-benh-tieu-duong-type-1-hay-type-2-nang-hon-4787888.html',
-         'https://vnexpress.net/bien-chung-ban-chan-tieu-duong-ma-khong-biet-4786682.html',
-         'https://vnexpress.net/6-mon-chong-viem-cho-nguoi-tieu-duong-4784664.html',
-         'https://vnexpress.net/phong-tranh-dai-thao-duong-cho-tre-4783833.html',
-         'https://vnexpress.net/dau-hieu-canh-bao-tang-duong-huyet-vao-ban-dem-4782698.html',
-         'https://vnexpress.net/dua-cai-chua-anh-huong-den-duong-huyet-the-nao-4780416.html',
-         'https://vnexpress.net/cach-phong-benh-tim-mach-do-tieu-duong-4779132.html',
-         'https://vnexpress.net/tai-sao-con-tre-van-mac-benh-dai-thao-duong-4778607.html',
-         'https://vnexpress.net/an-thit-lon-moi-ngay-co-tang-duong-huyet-4778290.html',
-         'https://vnexpress.net/7-yeu-to-loi-song-lam-tang-nguy-co-dai-thao-duong-4778159.html',
-         'https://vnexpress.net/bien-chung-dai-thao-duong-type-1-4775488.html',
-         'https://vnexpress.net/dau-hieu-tre-mac-benh-tieu-duong-type-1-4775097.html',
-         'https://vnexpress.net/vet-thuong-nao-canh-bao-bien-chung-ban-chan-tieu-duong-4773279.html',
-         'https://vnexpress.net/mat-thi-luc-bien-chung-am-tham-do-tieu-duong-4772949.html',
-         'https://vnexpress.net/loi-ich-cua-ca-rot-voi-nguoi-benh-tieu-duong-4772799.html',
-         'https://vnexpress.net/an-gao-lut-co-khoi-benh-dai-thao-duong-4771994.html',
-         'https://vnexpress.net/veo-chan-do-bien-chung-tieu-duong-ma-khong-biet-4771860.html',
-         'https://vnexpress.net/hai-phoi-trang-xoa-do-nhiem-trung-bien-chung-ban-chan-tieu-duong-4766234.html',
-         'https://vnexpress.net/nguoi-tieu-duong-nen-cao-voi-rang-may-lan-mot-nam-4763220.html',
-         'https://vnexpress.net/tai-sao-nguoi-tieu-duong-thuong-bi-hu-mong-tay-chan-4760998.html',
-         'https://vnexpress.net/nguyen-nhan-da-nguoi-tieu-duong-lao-hoa-nhanh-4760541.html',
-         'https://vnexpress.net/benh-vong-mac-tieu-duong-4759992.html',
-         'https://vnexpress.net/nhiem-trung-tu-vet-xuoc-nho-nguoi-benh-tieu-duong-phai-cat-bo-chan-4756894.html',
-         'https://vnexpress.net/ha-duong-huyet-do-hoi-chung-dumping-4754055.html',
-         'https://vnexpress.net/vet-xuoc-nho-tro-nang-do-bien-chung-tieu-duong-4752818.html',
-         'https://vnexpress.net/ha-duong-huyet-gay-dot-quy-gia-4749306.html',
-         'https://vnexpress.net/bat-ngo-phat-hien-benh-dai-thao-duong-khi-kham-vo-sinh-4738306.html',
-         'https://vnexpress.net/nsut-bang-thai-bi-hoai-tu-ban-chan-do-bien-chung-tieu-duong-4737469.html',
-         'https://vnexpress.net/hoai-tu-ngon-chan-do-khong-tuan-thu-dieu-tri-tieu-duong-4733422.html',
-         'https://vnexpress.net/mac-benh-tieu-duong-sau-5-nam-uong-thuoc-gia-truyen-4730328.html',
-         'https://vnexpress.net/lo-la-chua-tieu-duong-nguoi-phu-nu-gap-bien-chung-nang-4728271.html',
-         'https://vnexpress.net/tang-duong-huyet-4720398.html',
-         'https://vnexpress.net/vet-thuong-nhiem-trung-sau-ngam-vao-nuoc-muoi-4706944.html',
-         'https://vnexpress.net/tuong-het-tieu-duong-nguoi-phu-nu-bi-hoai-tu-chan-boi-mun-nhot-4706551.html',
-         'https://vnexpress.net/mun-nhot-nhiem-trung-gay-hoai-tu-mong-4705922.html',
-         'https://vnexpress.net/nhiem-trung-da-moi-biet-bi-tieu-duong-4704527.html',
-         'https://vnexpress.net/khoi-ap-xe-tu-co-lan-den-nguc-nguoi-dan-ong-4699712.html',
-         'https://vnexpress.net/vet-thuong-50-nam-kho-lanh-do-tieu-duong-4694776.html',
-         'https://vnexpress.net/hai-me-con-bi-bien-chung-tieu-duong-hoai-tu-chan-4688224.html',
-         'https://vnexpress.net/thieu-nu-15-tuoi-hon-me-bat-ngo-phat-hien-tieu-duong-4686751.html',
-         'https://vnexpress.net/suyt-mat-tinh-hoan-do-bien-chung-tieu-duong-4682982.html',
-         'https://vnexpress.net/viem-bao-quy-dau-do-bien-chung-tieu-duong-4670773.html',
-         'https://vnexpress.net/nhiem-trung-nang-do-bo-thuoc-tieu-duong-de-bam-huyet-4668235.html',
-         'https://vnexpress.net/xem-tieu-duong-nhu-cam-cum-nguoi-dan-ong-bien-chung-nang-4646698.html',
-         'https://vnexpress.net/bo-thuoc-chua-tieu-duong-nguoi-phu-nu-tang-duong-huyet-4643333.html',
-         'https://vnexpress.net/suyt-chet-sau-hai-thang-uong-thuoc-nam-chua-tieu-duong-4632037.html',
-         'https://vnexpress.net/nguy-kich-vi-thuong-xuyen-uong-nuoc-ngot-4608783.html',
-         'https://vnexpress.net/nguoi-benh-tieu-duong-suy-da-tang-do-con-trung-chich-4595219.html',
-         'https://vnexpress.net/nguy-kich-do-uong-ruou-bo-thuoc-tieu-duong-4590060.html',
-         'https://vnexpress.net/virus-tan-cong-vung-kin-nguoi-dai-thao-duong-gay-lo-loet-4584547.html',
-         'https://vnexpress.net/buon-non-do-cac-bien-chung-cua-benh-tieu-duong-4584593.html'
-'https://vnexpress.net/u-tai-do-thung-mang-nhi-4786434.html',
-                  'https://vnexpress.net/roi-loan-tien-dinh-co-the-chua-khoi-khong-4783817.html',
-                  'https://vnexpress.net/nhung-thoi-quen-gay-suy-giam-thinh-luc-4779660.html',
-                  'https://vnexpress.net/do-nhi-luong-de-lam-gi-4771858.html',
-                  'https://vnexpress.net/viem-tai-giua-4772425.html',
-                  'https://vnexpress.net/nhung-benh-thuong-gap-o-dai-tai-4782611.html',
-                  'https://vnexpress.net/hoa-mat-chong-mat-la-dau-hieu-roi-loan-tien-dinh-4753501.html',
-                  'https://vnexpress.net/chong-mat-tu-the-kich-phat-lanh-tinh-4748520.html',
-                  'https://vnexpress.net/nguy-co-giam-thinh-luc-do-deo-tai-nghe-am-luong-lon-4747235.html',
-                  'https://vnexpress.net/ret-chui-vao-tai-chang-trai-can-thung-mang-nhi-4725632.html',
-                  'https://vnexpress.net/nhieu-nguoi-viem-tai-sau-di-boi-giai-nhiet-4721376.html',
-                  'https://vnexpress.net/an-gi-giup-ban-nghe-tot-hon-4719908.html',
-                  'https://vnexpress.net/sai-lam-gay-hai-cho-tai-4714452.html',
-                  'https://vnexpress.net/thung-mang-nhi-boi-tieng-phao-no-4713105.html',
-                  'https://vnexpress.net/tre-viem-tai-giua-co-nen-di-may-bay-4708982.html',
-                  'https://vnexpress.net/meo-chong-say-tau-xe-ngay-tet-4708968.html',
-                  'https://vnexpress.net/deo-tai-nghe-co-giam-say-tau-xe-4708738.html',
-                  'https://vnexpress.net/bo-dieu-tri-roi-loan-tien-dinh-nang-4707644.html',
-                  'https://vnexpress.net/diec-mot-ben-tai-co-chua-khoi-duoc-khong-4706203.html',
-                  'https://vnexpress.net/nguyen-nhan-gay-tieng-keu-tach-tach-trong-tai-4703501.html',
-                  'https://vnexpress.net/loi-ich-khi-massage-tai-4703041.html',
-                  'https://vnexpress.net/nhung-dang-roi-loan-tien-dinh-thuong-gap-4702625.html',
-                  'https://vnexpress.net/roi-loan-tien-dinh-ma-khong-biet-4699428.html',
-                  'https://vnexpress.net/viem-tai-xuong-chum-4696311.html',
-                  'https://vnexpress.net/cach-giam-u-tai-khi-ngoi-may-bay-4694812.html',
-                  'https://vnexpress.net/tuong-quai-bi-hoa-u-tuyen-mang-tai-4691727.html',
-                  'https://vnexpress.net/ly-do-khong-nen-ngoay-tai-4690725.html',
-                  'https://vnexpress.net/nam-ong-tai-ngoai-co-tu-khoi-4685199.html',
-                  'https://vnexpress.net/tai-nghe-thanh-o-vi-khuan-khi-lau-ngay-khong-ve-sinh-4683244.html',
-                  'https://vnexpress.net/su-that-thu-vi-ve-tai-4682904.html',
-                  'https://vnexpress.net/viem-tai-giua-co-tu-khoi-khong-4679492.html',
-                  'https://vnexpress.net/nhieu-hat-cat-li-ti-trong-tai-be-trai-4673383.html',
-                  'https://vnexpress.net/nguyen-nhan-tai-chay-mau-4668710.html',
-                  'https://vnexpress.net/dau-hieu-bat-thuong-o-tai-canh-bao-benh-4667390.html',
-                  'https://vnexpress.net/nguyen-nhan-khien-tai-co-mui-kho-chiu-4667064.html',
-                  'https://vnexpress.net/tieng-on-anh-huong-den-tai-the-nao-4665938.html',
-                  'https://vnexpress.net/bac-si-viet-va-maylaysia-trinh-dien-cay-oc-tai-benh-nhan-4664853.html',
-                  'https://vnexpress.net/6-thuc-pham-khong-tot-cho-nguoi-roi-loan-tien-dinh-4662012.html',
-                  'https://vnexpress.net/lam-gi-khi-bi-roi-loan-tien-dinh-4660749.html',
-                  'https://vnexpress.net/bao-ve-tai-nhu-the-nao-4658769.html',
-                  'https://vnexpress.net/nguyen-nhan-u-tai-keo-dai-4658171.html',
-                  'https://vnexpress.net/viem-tai-giua-lam-thung-mang-nhi-4657600.html',
-                  'https://vnexpress.net/an-gi-giam-roi-loan-tien-dinh-4657007.html',
-                  'https://vnexpress.net/sai-lam-khi-dung-tai-nghe-4656614.html',
-                  'https://vnexpress.net/thung-mang-nhi-boi-chiec-tam-bong-ngoay-tai-4655936.html',
-                  'https://vnexpress.net/doan-benh-qua-mau-ray-tai-4653682.html',
-                  'https://vnexpress.net/nguyen-nhan-tich-tu-ray-tai-4652578.html',
-                  'https://vnexpress.net/vi-sao-khong-nen-deo-tai-nghe-khi-ngu-4646592.html',
-                  'https://vnexpress.net/nut-ray-4-cm-bit-kin-hai-tai-be-trai-4646171.html',
-                  'https://vnexpress.net/viem-tai-xuong-chum-4643833.html',
-                  'https://vnexpress.net/nhiem-trung-tai-do-nam-4641430.html',
-                  'https://vnexpress.net/6-cach-giam-tac-tai-4640485.html',
-                  'https://vnexpress.net/chong-mat-buon-non-trieu-chung-roi-loan-tien-dinh-4636237.html',
-                  'https://vnexpress.net/khi-nao-can-tham-kham-nhiem-trung-tai-4628742.html',
-                  'https://vnexpress.net/thung-mang-nhi-vi-nhiem-nam-ong-tai-4620633.html',
-                  'https://vnexpress.net/roi-loan-tien-dinh-nen-an-gi-kieng-gi-4620044.html',
-                  'https://vnexpress.net/nguyen-nhan-gay-dau-ong-tai-4618315.html',
-                  'https://vnexpress.net/cac-nguyen-nhan-gay-sung-tai-va-cach-dieu-tri-4616066.html',
-                  'https://vnexpress.net/9-nguyen-nhan-gay-vay-trong-tai-4607599.html',
-                  'https://vnexpress.net/xu-tri-the-nao-khi-kien-chui-vao-tai-4605776.html',
-                  'https://vnexpress.net/5-bai-tap-cai-thien-chong-mat-do-roi-loan-tien-dinh-4599505.html',
-                  'https://vnexpress.net/4-meo-giup-ngu-ngon-khi-bi-u-tai-4597235.html',
-                  'https://vnexpress.net/cac-bien-chung-nguy-hiem-cua-nhiem-trung-tai-4595383.html',
-                  'https://vnexpress.net/roi-loan-tien-dinh-co-nen-uong-thuoc-bo-nao-4593772.html',
-                  'https://vnexpress.net/roi-loan-tien-dinh-co-gay-dot-quy-khong-4592020.html',
-                  'https://vnexpress.net/tac-hai-khi-nhin-hat-hoi-4591683.html',
-                  'https://vnexpress.net/cac-nguyen-nhan-gay-dau-tai-thuong-gap-4590884.html',
-                  'https://vnexpress.net/5-benh-ly-khien-tai-dau-nhoi-4590553.html',
-                  'https://vnexpress.net/ly-do-nghe-thay-tieng-mach-dap-trong-tai-4586987.html',
-                  'https://vnexpress.net/nhung-phuong-phap-loai-bo-ray-tai-4585706.html',
-                  'https://vnexpress.net/nhung-dong-tac-yoga-giup-giam-cac-benh-ve-tai-4585485.html',
-                  'https://vnexpress.net/nhan-biet-chong-mat-do-roi-loan-tien-dinh-4584998.html',
-                  'https://vnexpress.net/4-bien-chung-do-xo-lo-tai-4582203.html',
-                  'https://vnexpress.net/nguyen-nhan-tai-bi-chay-mau-4581182.html',
-                  'https://vnexpress.net/moi-nguy-khi-lay-ray-tai-bang-tam-bong-4574824.html',
-                  'https://vnexpress.net/60-nguyen-nhan-gay-diec-bam-sinh-co-the-phong-ngua-4572719.html',
-                  'https://vnexpress.net/khi-nao-can-dung-may-tro-thinh-4571769.html',
-                  'https://vnexpress.net/dieu-nen-lam-nen-tranh-khi-chua-nhiem-trung-tai-4570561.html',
-'https://vnexpress.net/4-benh-mui-xoang-de-nham-lan-trieu-chung-4782494.html',
-                  'https://vnexpress.net/lech-vach-ngan-mui-co-tai-phat-sau-phau-thuat-4780752.html',
-                  'https://vnexpress.net/mat-khuu-giac-co-the-chua-khoi-khong-4779811.html',
-                  'https://vnexpress.net/benh-ngu-ngay-va-cach-dieu-tri-4776375.html',
-                  'https://vnexpress.net/9-cach-don-gian-giup-giam-so-mui-4774663.html',
-                  'https://vnexpress.net/nhung-hieu-lam-thuong-gap-ve-viem-xoang-4767610.html',
-                  'https://vnexpress.net/cuon-mui-gian-no-do-lam-dung-thuoc-nho-mui-4767601.html',
-                  'https://vnexpress.net/tre-tai-phat-viem-mui-xoang-do-di-boi-4765900.html',
-                  'https://vnexpress.net/cham-soc-sau-phau-thuat-lech-vach-ngan-mui-the-nao-4765588.html',
-                  'https://vnexpress.net/viem-mui-xoang-di-ung-co-can-xet-nghiem-tim-nguyen-nhan-4763199.html',
-                  'https://vnexpress.net/lam-dung-thuoc-nho-mui-gay-bien-chung-4756694.html',
-                  'https://vnexpress.net/tuong-u-tai-hoa-phi-dai-cuon-mui-4753783.html',
-                  'https://vnexpress.net/qua-phat-cuon-mui-4753159.html'
-                  'https://vnexpress.net/nen-va-khong-nen-lam-gi-khi-cam-lanh-4722959.html',
-                  'https://vnexpress.net/nhieu-nguoi-chay-mau-cam-do-nang-nong-4718663.html',
-                  'https://vnexpress.net/nhieu-nguoi-viem-tai-mui-hong-sau-tet-4716515.html',
-                  'https://vnexpress.net/viem-mui-di-ung-mua-xuan-4711777.html',
-                  'https://vnexpress.net/khoi-huong-anh-huong-den-nguoi-viem-mui-xoang-the-nao-4709688.html',
-                  'https://vnexpress.net/co-nen-phau-thuat-chua-phi-dai-cuon-mui-4707483.html',
-                  'https://vnexpress.net/trieu-chung-viem-mui-xoang-4707387.html',
-                  'https://vnexpress.net/he-qua-kho-luong-khi-nhin-hat-hoi-4705705.html',
-                  'https://vnexpress.net/8-cach-giup-giam-di-ung-4697348.html',
-                  'https://vnexpress.net/tai-sao-ban-chay-nuoc-mui-4697355.html',
-                  'https://vnexpress.net/dau-hieu-nhan-biet-ung-thu-mui-4694235.html',
-                  'https://vnexpress.net/cach-ngan-ngua-viem-mui-di-ung-4692565.html',
-                  'https://vnexpress.net/viem-xoang-chay-mau-mui-nguy-hiem-khong-4692049.html',
-                  'https://vnexpress.net/dau-hieu-nhan-biet-ung-thu-xoang-4691189.html',
-                  'https://vnexpress.net/cach-tranh-di-ung-voi-cay-thong-noel-4690202.html',
-                  'https://vnexpress.net/su-that-thu-vi-ve-mui-nguoi-4688828.html',
-                  'https://vnexpress.net/nguyen-nhan-khuu-giac-cua-ban-thay-doi-4688253.html',
-                  'https://vnexpress.net/cach-don-gian-phong-viem-xoang-4687834.html',
-                  'https://vnexpress.net/mo-xoang-bao-lau-hoi-phuc-4683467.html',
-                  'https://vnexpress.net/an-gi-khi-chay-mau-cam-4680718.html',
-                  'https://vnexpress.net/dau-dau-do-viem-xoang-dieu-tri-the-nao-4680241.html',
-                  'https://vnexpress.net/trieu-chung-nhiem-trung-xoang-lan-len-nao-4676904.html',
-                  'https://vnexpress.net/hep-eo-hong-nguoi-dan-ong-thuong-ngat-tho-khi-ngu-4675560.html',
-                  'https://vnexpress.net/tai-sao-mui-thuong-nghet-vao-buoi-sang-4674149.html',
-                  'https://vnexpress.net/nam-lap-day-cac-xoang-cua-nguoi-dan-ong-4671435.html',
-                  'https://vnexpress.net/polyp-mui-co-tu-khoi-khong-4670862.html',
-                  'https://vnexpress.net/7-thuc-pham-tot-cho-nguoi-benh-viem-xoang-4670701.html',
-                  'https://vnexpress.net/nguyen-nhan-kho-mui-4670163.html',
-                  'https://vnexpress.net/5-loai-tra-giup-giam-nghet-mui-4667503.html',
-                  'https://vnexpress.net/dau-hieu-nhiem-ky-sinh-trung-tai-mui-hong-4559846.html',
-                  'https://vnexpress.net/viem-xoang-co-lay-khi-thoi-tiet-lanh-4550783.html',
-                  'https://vnexpress.net/8-trieu-chung-canh-bao-mac-nhiem-trung-xoang-4548151.html',
-                  'https://vnexpress.net/viem-xoang-do-nam-4545696.html',
-                  'https://vnexpress.net/cach-xu-tri-viem-xoang-tai-phat-mua-lanh-4544208.html',
-                  'https://vnexpress.net/chay-mau-cam-do-di-ung-4542679.html',
-                  'https://vnexpress.net/cac-bai-tap-tho-tot-cho-mui-4542459.html',
-                  'https://vnexpress.net/xu-tri-nghet-mui-cho-tre-dung-cach-4541909.html',
-                  'https://vnexpress.net/5-sai-lam-can-tranh-khi-tu-dieu-tri-nhiem-trung-xoang-4539447.html',
-                  'https://vnexpress.net/4-loai-thuc-pham-nguoi-benh-viem-xoang-nen-han-che-4538012.html',
-                  'https://vnexpress.net/cach-xu-ly-mun-moc-o-mui-4537368.html',
-                  'https://vnexpress.net/cach-xu-ly-mun-moc-o-mui-4537368.html',
-                  'https://vnexpress.net/dau-khi-nhai-canh-bao-sau-rang-ung-thu-mieng-4537269.html',
-                  'https://vnexpress.net/cach-cai-thien-di-ung-mui-4534791.html',
-                  'https://vnexpress.net/hut-thuoc-la-gay-hai-cho-mui-xoang-4534212.html',
-                  'https://vnexpress.net/co-nen-dung-nuoc-muoi-rua-mui-cho-tre-keo-dai-4533758.html',
-                  'https://vnexpress.net/viem-mui-khi-mang-thai-4532926.html',
-                  'https://vnexpress.net/8-nguyen-nhan-gay-ngua-mui-4530918.html',
-                  'https://vnexpress.net/nguyen-nhan-gay-chay-dich-mui-sau-4513778.html',
-                  'https://vnexpress.net/8-meo-giup-de-ngu-khi-bi-nghet-mui-4512537.html',
-                  'https://vnexpress.net/5-cach-giup-giam-so-mui-4511128.html',
-                  'https://vnexpress.net/8-meo-giup-cai-thien-ngua-mui-4506675.html',
-                  'https://vnexpress.net/4-cach-giam-trieu-chung-viem-xoang-tai-nha-4506258.html',
-                  'https://vnexpress.net/chay-mau-cam-do-cang-thang-4506556.html',
-                  'https://vnexpress.net/cach-giam-nghet-mui-don-gian-tai-nha-4501045.html',
-                  'https://vnexpress.net/cach-xong-hoi-bang-thao-moc-giup-thong-mui-4496788.html',
-'https://vnexpress.net/noi-hach-o-co-co-phai-ung-thu-4780340.html',
-                  'https://vnexpress.net/ai-co-nguy-co-ung-thu-vom-hong-4764545.html',
-                  'https://vnexpress.net/khi-nao-nen-tam-soat-ung-thu-tai-mui-hong-4756286.html',
-                  'https://vnexpress.net/u-tuyen-nuoc-bot-4754883.html',
-                  'https://vnexpress.net/viem-mui-hong-thuong-xuyen-do-va-khong-tieu-bien-4777385.html',
-                  'https://vnexpress.net/ung-thu-tai-mui-hong-4728889.html',
-                  'https://vnexpress.net/ung-thu-ham-4726086.html',
-                  'https://vnexpress.net/ung-thu-xoang-4724895.html',
-                  'https://vnexpress.net/ung-thu-amidan-4717052.html',
-                  'https://vnexpress.net/co-can-phau-thuat-soi-tuyen-nuoc-bot-duoi-luoi-4712326.html',
-                  'https://vnexpress.net/ly-giai-ve-phan-xa-ngap-4682425.html',
-                  'https://vnexpress.net/5-mon-an-phong-ung-thu-dau-co-4664707.html',
-                  'https://vnexpress.net/cach-dao-thai-soi-tuyen-nuoc-bot-4639958.html'
-                  'https://vnexpress.net/bien-dang-mat-do-soi-duoi-ham-4638478.html',
-                  'https://vnexpress.net/mu-mat-trai-sau-tai-nan-chan-thuong-ham-mat-4601708.html',
-                  'https://vnexpress.net/chua-tram-cam-mot-nam-moi-biet-roi-loan-tien-dinh-4541714.html',
-                  'https://vnexpress.net/nhai-nuot-kho-di-kham-phat-hien-u-xuong-ham-4527576.html',
-                  'https://vnexpress.net/cac-loai-thuc-uong-giup-giam-nhiet-mieng-4489665.html',
-                  'https://vnexpress.net/7-dau-hieu-canh-bao-ung-thu-thuc-quan-4487585.html',
-                  'https://vnexpress.net/dau-hieu-canh-bao-ung-thu-vom-hong-4481963.html',
-                  'https://vnexpress.net/hon-co-the-lam-lay-lan-nhung-benh-gi-4455894.html',
-                  'https://vnexpress.net/oral-sex-co-the-lay-truyen-virus-gay-ung-thu-vom-hong-4455396.html',
-                  'https://vnexpress.net/9-thuc-pham-giau-iot-phong-ngua-buou-co-4455057.html',
-                  'https://vnexpress.net/noi-hach-o-co-khi-nao-nguy-hiem-4446139.html',
-                  'https://vnexpress.net/5-loai-ung-thu-vung-dau-co-thuong-gap-4446226.html',
-                  'https://vnexpress.net/dieu-tri-u-nang-keo-tuyen-giap-4400886.html',
-                  'https://vnexpress.net/u-tuyen-giap-khi-nao-phat-trien-thanh-ung-thu-4398938.html'
-'https://vnexpress.net/dieu-tri-roi-loan-giong-noi-tuoi-day-thi-the-nao-4781180.html',
-                  'https://vnexpress.net/ve-sinh-rang-mieng-ky-sao-hoi-tho-van-co-mui-4778484.html',
-                  'https://vnexpress.net/nam-noi-giong-nu-chua-nhu-the-nao-4778128.html',
-                  'https://vnexpress.net/benh-ngu-ngay-va-cach-dieu-tri-4776375.html',
-                  'https://vnexpress.net/viem-mui-hong-thuong-xuyen-do-va-khong-tieu-bien-4777385.html',
-                  'https://vnexpress.net/tho-khi-dung-tai-nha-duoc-khong-4775466.html',
-                  'https://vnexpress.net/vet-loet-do-nhiet-mieng-co-khac-ung-thu-luoi-4775220.html',
-                  'https://vnexpress.net/cac-mon-nen-an-va-tranh-khi-nhiet-mieng-4774741.html',
-                  'https://vnexpress.net/cham-soc-sau-phau-thuat-lech-vach-ngan-mui-the-nao-4765588.html',
-                  'https://vnexpress.net/nhung-benh-thuong-gay-dau-hong-4771957.html',
-                  'https://vnexpress.net/hoi-chung-ngung-tho-khi-ngu-4770417.html',
-                  'https://vnexpress.net/co-nen-dung-nuoc-muoi-sinh-ly-suc-hong-rua-mui-moi-ngay-4768947.html',
-                  'https://vnexpress.net/co-can-cat-thang-luoi-cho-tre-4768561.html'
-                  'https://vnexpress.net/manh-xuong-ca-dam-vao-luoi-nguoi-dan-ong-4766366.html',
-                  'https://vnexpress.net/meo-phong-dau-hong-ngay-nang-4765450.html',
-                  'https://vnexpress.net/nguoi-lon-tuoi-co-phau-thuat-amidan-duoc-khong-4763918.html',
-                  'https://vnexpress.net/ngu-ngay-co-phai-la-benh-4763644.html',
-                  'https://vnexpress.net/cat-amidan-bao-lau-co-the-an-uong-binh-thuong-4761724.html',
-                  'https://vnexpress.net/hoc-di-vat-mui-hong-tai-nan-thuong-gap-o-tre-nho-4760859.html',
-                  'https://vnexpress.net/viem-amidan-benh-de-tai-phat-mua-he-4758338.html',
-                  'https://vnexpress.net/suc-khoe/cac-benh/benh-tai-mui-hong/hong-thanh-quan-p4',
-                  'https://vnexpress.net/mat-giong-do-polyp-day-thanh-quan-4751454.html',
-                  'https://vnexpress.net/hoi-chung-bong-rat-mieng-4727520.html',
-                  'https://vnexpress.net/5-o-chua-nhieu-mam-benh-cam-lanh-4723569.html',
-                  'https://vnexpress.net/viem-amidan-hoc-mu-4719193.html',
-                  'https://vnexpress.net/9-mon-an-uong-tot-cho-nguoi-dau-hong-4718648.html',
-                  'https://vnexpress.net/mau-sac-o-luoi-canh-bao-benh-gi-4718257.html',
-                  'https://vnexpress.net/9-benh-ve-luoi-4715719.html',
-                  'https://vnexpress.net/cat-amidan-bao-lau-co-the-uong-ruou-bia-4709682.html',
-                  'https://vnexpress.net/cach-nao-phong-nhiet-mieng-ngay-tet-4710048.html',
-                  'https://vnexpress.net/cach-giam-kho-rat-hong-sau-khi-uong-ruou-bia-4709946.html',
-                  'https://vnexpress.net/nguyen-nhan-gay-benh-luoi-trang-4708565.html',
-                  'https://vnexpress.net/u-tuyen-nuoc-bot-co-thanh-ung-thu-4706609.html',
-                  'https://vnexpress.net/u-tuyen-nuoc-bot-duoi-ham-4705269.html',
-                  'https://vnexpress.net/dau-hieu-o-luoi-canh-bao-co-the-thieu-vitamin-d-4704352.html',
-                  'https://vnexpress.net/tai-sao-ngay-to-hon-sau-khi-uong-ruou-4704398.html',
-                  'https://vnexpress.net/meo-giam-kho-mieng-4702038.html',
-                  'https://vnexpress.net/tai-sao-chay-nuoc-dai-khi-ngu-4700805.html',
-                  'https://vnexpress.net/tai-sao-chay-nuoc-dai-khi-ngu-4700805.html',
-                  'https://vnexpress.net/khan-tieng-canh-bao-benh-gi-4700518.html',
-                  'https://vnexpress.net/sai-lam-khi-dung-nuoc-suc-mieng-4699734.html',
-                  'https://vnexpress.net/nguyen-nhan-hoi-tho-co-mui-khong-do-thuc-pham-4699028.html',
-                  'https://vnexpress.net/9-nguyen-nhan-gay-ngu-ngay-4698222.html',
-                  'https://vnexpress.net/tai-tao-khi-quan-cho-nguoi-bi-bien-chung-sau-dot-quy-4696683.html',
-                  'https://vnexpress.net/ho-khien-co-the-dot-chay-bao-nhieu-nang-luong-4695941.html',
-                  'https://vnexpress.net/6-cach-lam-diu-co-hong-sau-khi-non-4694508.html',
-                  'https://vnexpress.net/6-mon-an-uong-lam-tang-dom-4694925.html',
-                  'https://vnexpress.net/6-mon-an-uong-tieu-dom-4693471.html',
-                  'https://vnexpress.net/7-cach-loai-bo-dom-o-hong-4692983.html',
-                  'https://vnexpress.net/6-mon-an-tang-mien-dich-phong-cam-cum-4689193.html',
-                  'https://vnexpress.net/cach-giam-dau-hong-co-dom-mua-lanh-4687824.html',
-                  'https://vnexpress.net/cam-giac-on-lanh-do-dau-4687552.html',
-                  'https://vnexpress.net/phau-thuat-3-trong-1-chua-ngu-ngay-4684271.html',
-                  'https://vnexpress.net/meo-bao-ve-giong-noi-trong-treo-4683942.html',
-                  'https://vnexpress.net/mo-tuyen-giap-lac-cho-khien-nguoi-phu-nu-kho-nuot-4683315.html',
-                  'https://vnexpress.net/tai-sao-kho-mieng-4680250.html',
-                  'https://vnexpress.net/kieng-gi-khi-viem-hong-de-nhanh-khoi-benh-4678212.html',
-                  'https://vnexpress.net/nguyen-nhan-gay-chua-mieng-4675373.html',
-                  'https://vnexpress.net/cong-dung-giam-ho-cua-tinh-dau-4672944.html',
-                  'https://vnexpress.net/viem-amidan-man-tinh-co-gay-ung-thu-4671019.html',
-                  'https://vnexpress.net/ngu-ngay-nhu-sam-bat-ngo-phat-hien-viem-amidan-4668372.html',
-                  'https://vnexpress.net/6-cach-giam-ngua-hong-4650860.html',
-'https://vnexpress.net/5-mon-an-uong-giup-ha-men-gan-tu-nhien-4783142.html',
-        'https://vnexpress.net/nhung-dieu-nen-lam-de-bao-ve-gan-4783185.html',
-        'https://vnexpress.net/4-nhom-thuc-pham-tot-cho-nguoi-viem-tuy-4782239.html',
-        'https://vnexpress.net/noi-me-day-co-phai-dau-hieu-suy-giam-chuc-nang-gan-4781330.html',
-        'https://vnexpress.net/noi-me-day-co-phai-dau-hieu-suy-giam-chuc-nang-gan-4781330.html',
-        'https://vnexpress.net/hang-chuc-polyp-bam-nhu-thach-nhu-trong-tui-mat-4781230.html',
-        'https://vnexpress.net/vitamin-nao-tot-cho-gan-4780108.html',
-        'https://vnexpress.net/loi-ich-cua-dau-nanh-voi-gan-4778700.html',
-        'https://vnexpress.net/loi-ich-cua-dau-nanh-voi-gan-4778700.html',
-        'https://vnexpress.net/bi-xo-gan-an-uong-the-nao-4778210.html',
-        'https://vnexpress.net/4-phuong-phap-dieu-tri-viem-tuy-cap-4775464.html',
-        'https://vnexpress.net/uong-ruou-lam-ton-thuong-gan-the-nao-4775123.html',
-        'https://vnexpress.net/an-gung-co-tot-cho-gan-4774650.html',
-        'https://vnexpress.net/an-gung-co-tot-cho-gan-4774650.html',
-        'https://vnexpress.net/nhung-dieu-nen-lam-de-bao-ve-gan-4783185.html',
-        'https://vnexpress.net/4-nhom-thuc-pham-tot-cho-nguoi-viem-tuy-4782239.html',
-        'https://vnexpress.net/5-thoi-quen-xau-lam-tang-nguy-co-viem-tuy-cap-4781596.html',
-        'https://vnexpress.net/khoi-nguy-co-ghep-gan-nho-thay-24-lit-huyet-tuong-4779669.html',
-        'https://vnexpress.net/vitamin-nao-tot-cho-gan-4780108.html',
-        'https://vnexpress.net/5-loai-rau-cu-qua-ho-tro-ha-men-gan-4780481.html',
-        'https://vnexpress.net/4-nhom-thuc-pham-nguoi-men-gan-cao-nen-an-4769537.html',
-        'https://vnexpress.net/uong-ca-phe-loi-hay-hai-gan-4769129.html',
-        'https://vnexpress.net/yeu-to-nguy-co-lam-tang-kich-thuoc-soi-tui-mat-4763637.html',
-        'https://vnexpress.net/5-bo-phan-sung-tay-bat-thuong-canh-bao-benh-gan-4763456.html',
-        'https://vnexpress.net/an-uong-gi-giai-doc-gan-tu-nhien-4760588.html',
-        'https://vnexpress.net/thieu-ngu-lam-ton-thuong-gan-4759637.html',
-        'https://vnexpress.net/benh-nao-gan-4759539.html',
-        'https://vnexpress.net/4-loai-sua-tot-cho-nguoi-benh-gan-nhiem-mo-4758745.html',
-        'https://vnexpress.net/nguoi-benh-xo-gan-an-thit-bo-duoc-khong-4756838.html',
-        'https://vnexpress.net/lam-the-nao-ha-men-gan-4756824.html',
-        'https://vnexpress.net/8-thoi-quen-giup-phong-benh-gan-4756421.html',
-        'https://vnexpress.net/9-cach-giup-giam-nguy-co-mac-benh-tui-mat-4755656.html',
-        'https://vnexpress.net/an-man-hai-suc-khoe-the-nao-4755900.html',
-        'https://vnexpress.net/thu-pham-gay-tang-men-gan-4754439.html',
-        'https://vnexpress.net/soi-tuy-co-nguy-hiem-khong-4754882.html',
-        'https://vnexpress.net/nhiem-trung-duong-mat-do-soi-bit-kin-ong-gan-4755260.html',
- 'https://vnexpress.net/kham-suc-khoe-hoc-duong-bac-si-phat-hien-be-gai-bat-thuong-tim-phoi-4688876.html',
-        'https://vnexpress.net/vi-khuan-an-thit-nguoi-co-lay-khong-4535888.html',
-        'https://vnexpress.net/viem-phoi-do-nam-4520092.html',
-        'https://vnexpress.net/dau-hieu-nhan-biet-tre-co-trieu-chung-ho-hap-hau-covid-19-4452703.html',
-        'https://vnexpress.net/nhung-dieu-can-biet-ve-thuoc-dieu-tri-hen-suyen-4437165.html',
-        'https://vnexpress.net/benh-bui-phoi-silic-pho-bien-o-nguoi-tren-40-4436530.html',
-        'https://vnexpress.net/benh-sarcoidosis-dau-hieu-nguyen-nhan-va-phuong-phap-dieu-tri-4436531.html',
-        'https://vnexpress.net/benh-phoi-hiem-gap-khien-be-gai-9-nam-thieu-mau-3730384.html',
-        'https://vnexpress.net/kho-tho-do-vo-ken-khi-phoi-4769681.html',
-'https://vnexpress.net/5-thuc-pham-giup-giam-ho-4787558.html',
-        'https://vnexpress.net/che-do-dinh-duong-cho-tre-viem-duong-ho-hap-4787582.html',
-        'https://vnexpress.net/4-luu-y-de-tho-dung-cach-khi-chay-bo-4787037.html',
-        'https://vnexpress.net/thu-pham-gay-benh-ho-hap-o-nguoi-cao-tuoi-khi-giao-mua-4786497.html',
-        'https://vnexpress.net/5-cach-ngan-mui-hoi-tho-buoi-sang-4786394.html',
-        'https://vnexpress.net/phoi-ton-thuong-hinh-choi-cay-do-lao-4786367.html',
-        'https://vnexpress.net/nhung-mon-an-uong-nen-tranh-khi-ho-4786135.html',
-        'https://vnexpress.net/nguyen-nhan-it-ngo-gay-kho-tho-4785689.html',
-        'https://vnexpress.net/7-loai-thao-moc-giup-giam-mui-hoi-tho-4785295.html',
-        'https://vnexpress.net/lam-the-nao-giam-ho-nhanh-khong-dung-thuoc-4785261.html',
-        'https://vnexpress.net/cam-cum-va-cam-lanh-khac-nhau-the-nao-4784903.html',
-        'https://vnexpress.net/lam-the-nao-khu-mui-bia-ruou-gay-hoi-mieng-4784787.html',
-        'https://vnexpress.net/duong-lay-nhiem-vi-khuan-an-thit-nguoi-4784108.html',
-        'https://vnexpress.net/an-uong-the-nao-de-hoi-tho-thom-mat-4783505.html',
-        'https://vnexpress.net/ly-do-phu-nu-man-kinh-thuong-ngay-ngu-4783577.html',
-        'https://vnexpress.net/ban-biet-gi-ve-tho-5-ngon-tay-4783181.html',
-        'https://vnexpress.net/dung-la-trau-the-nao-giam-mui-hoi-mieng-4782713.html',
-        'https://vnexpress.net/lam-the-nao-phong-benh-ho-hap-cho-tre-mua-tuu-truong-4782411.html',
-        'https://vnexpress.net/nhung-loai-cu-qua-nen-an-thuong-xuyen-de-tang-mien-dich-4782034.html',
-        'https://vnexpress.net/4-thoi-quen-an-uong-khien-hoi-tho-co-mui-4781725.html',
-        'https://vnexpress.net/ngu-ngay-gay-hai-the-nao-4781321.html',
-        'https://vnexpress.net/gia-vi-nao-ho-tro-thong-thoang-duong-tho-4781039.html',
-        'https://vnexpress.net/nguyen-nhan-khien-phoi-lao-hoa-4780476.html',
-        'https://vnexpress.net/bai-tap-nao-giup-giam-ngu-ngay-4780112.html',
-        'https://vnexpress.net/thoi-quen-truoc-khi-ngu-tot-cho-ho-hap-4779874.html',
-        'https://vnexpress.net/5-lam-tuong-ve-ngung-tho-khi-ngu-4779493.html',
-        'https://vnexpress.net/thieu-ngu-tang-nguy-co-mac-benh-ho-hap-4779193.html',
-        'https://vnexpress.net/bi-quyet-chay-bo-khong-kho-tho-4779096.html',
-        'https://vnexpress.net/thoi-quen-nao-khien-phoi-dan-suy-kiet-4778737.html',
-        'https://vnexpress.net/ve-sinh-phong-ngu-dung-cach-the-nao-4777867.html',
-        'https://vnexpress.net/tran-dich-mang-phoi-4777406.html',
-        'https://vnexpress.net/nghien-thuoc-la-bia-ruou-tang-nguy-co-ngung-tho-khi-ngu-4776181.html',
-        'https://vnexpress.net/benh-lao-4775730.html',
-        'https://vnexpress.net/dau-hieu-thieu-oxy-khi-ngu-4775720.html',
-        'https://vnexpress.net/thuyen-tac-phoi-4773702.html',
-        'https://vnexpress.net/10-cach-tang-kha-nang-mien-dich-cua-co-the-4773738.html',
-        'https://vnexpress.net/cach-don-gian-thanh-loc-khong-khi-trong-nha-4773281.html',
-        'https://vnexpress.net/an-gi-de-tang-mien-dich-it-om-vat-4773000.html',
-        'https://vnexpress.net/khoi-bep-an-co-the-gay-benh-phoi-4772286.html',
-        'https://vnexpress.net/an-nhieu-kem-uong-nuoc-da-de-benh-ho-hap-4771666.html',
-        'https://vnexpress.net/trieu-chung-benh-bach-hau-khac-cam-covid-19-the-nao-4769080.html',
-        'https://vnexpress.net/trieu-chung-benh-bach-hau-khac-cam-covid-19-the-nao-4769080.html',
-        'https://vnexpress.net/5-thuc-pham-giup-giam-ngu-ngay-4769152.html',
-        'https://vnexpress.net/cac-mon-an-co-the-hai-phoi-4768356.html',
-        'https://vnexpress.net/6-mon-an-uong-bo-phoi-4766949.html',
-        'https://vnexpress.net/nguoi-benh-hen-suyen-co-nen-uong-ca-phe-muoi-4766842.html',
-        'https://vnexpress.net/nhung-dieu-it-biet-ve-phoi-4761992.html',
-        'https://vnexpress.net/viem-phoi-do-nhiem-nam-4761139.html',
-        'https://vnexpress.net/dau-hieu-nao-canh-bao-can-kham-phoi-4758672.html',
-        'https://vnexpress.net/mang-phoi-san-sui-nhu-san-ho-do-ung-thu-di-can-4758010.html',
-        'https://vnexpress.net/cham-soc-tre-so-sinh-viem-phoi-4757725.html',
-        'https://vnexpress.net/tai-sao-phoi-suy-yeu-4755651.html',
-        'https://vnexpress.net/tai-sao-viem-hong-de-bien-thanh-viem-phoi-4751230.html',
-        'https://vnexpress.net/ngung-tho-khi-ngu-4750578.html',
-        'https://vnexpress.net/lay-cum-trong-gia-dinh-ba-nguoi-nhap-vien-4750166.html',
-        'https://vnexpress.net/hut-thuoc-la-hai-phoi-the-nao-4743290.html',
-        'https://vnexpress.net/khong-hut-thuoc-nhieu-nguoi-van-mac-ung-thu-phoi-4737842.html',
-        'https://vnexpress.net/nhieu-nguoi-nhap-vien-vi-cum-mua-bien-chung-viem-phoi-4737858.html',
-        'https://vnexpress.net/kiem-tra-de-biet-ban-co-nguy-co-ung-thu-phoi-khong-4735062.html',
-        'https://vnexpress.net/sot-ret-keo-dai-phat-hien-lao-mang-phoi-4732738.html',
-        'https://vnexpress.net/cu-ong-nam-mot-cho-lau-ngay-bi-viem-phoi-loet-da-nang-4730575.html',
-        'https://vnexpress.net/bien-chung-nguy-hiem-cua-cum-4726260.html',
-        'https://vnexpress.net/ai-co-nguy-co-mac-benh-bui-phoi-4725881.html',
-        'https://vnexpress.net/mac-thuy-dau-bien-chung-thanh-viem-phoi-4725419.html',
-        'https://vnexpress.net/ho-ra-mau-o-at-do-gian-dong-mach-phe-quan-4724513.html',
-        'https://vnexpress.net/mui-nen-thom-co-the-gay-benh-4719117.html',
-        'https://vnexpress.net/tang-ap-phoi-tro-nang-khien-nguoi-phu-nu-ngat-xiu-4718196.html',
-        'https://vnexpress.net/ap-xe-phoi-do-hoc-xuong-4715356.html',
-        'https://vnexpress.net/dau-hieu-nhiem-virus-hop-bao-ho-hap-4714592.html',
-        'https://vnexpress.net/viem-trang-phoi-sau-chuyen-leo-nui-4708745.html'
-'https://vnexpress.net/ba-phuong-phap-dieu-tri-suy-tuyen-yen-4787721.html',
-        'https://vnexpress.net/nhan-giap-xam-lan-sun-khi-quan-nguoi-dan-ong-4787623.html',
-        'https://vnexpress.net/dau-hieu-tre-thieu-hormone-tang-truong-4782224.html',
-        'https://vnexpress.net/dung-uong-sua-bo-co-ngan-day-thi-som-4777030.html',
-        'https://vnexpress.net/mo-tham-my-nhan-tuyen-giap-4776624.html',
-        'https://vnexpress.net/ha-canxi-mau-4771495.html',
-        'https://vnexpress.net/tiem-hormone-kim-ham-day-thi-som-cho-be-gai-7-tuoi-4765899.html',
-        'https://vnexpress.net/dai-thao-nhat-trung-uong-4764597.html',
-        'https://vnexpress.net/suy-tuyen-thuong-than-la-benh-gi-4764233.html',
-        'https://vnexpress.net/dai-thao-nhat-do-than-4752030.html',
-        'https://vnexpress.net/tuong-benh-duong-ruot-hoa-suy-giap-4750597.html',
-        'https://vnexpress.net/ha-kali-mau-4749619.html',
-        'https://vnexpress.net/7-mon-an-bo-sung-noi-tiet-to-nu-4749226.html',
-        'https://vnexpress.net/tang-huyet-ap-keo-dai-moi-phat-hien-do-u-tuyen-thuong-than-4747608.html',
-        'https://vnexpress.net/tu-choi-mo-u-tuy-thuong-than-nguoi-phu-nu-nhieu-lan-tang-huyet-ap-4745304.html',
-        'https://vnexpress.net/10-tu-the-yoga-giup-can-bang-noi-tiet-to-4743714.html',
-        'https://vnexpress.net/loi-ich-cua-yoga-trong-dieu-hoa-noi-tiet-to-4743530.html',
-        'https://vnexpress.net/meo-can-bang-noi-tiet-to-nu-4739829.html',
-        'https://vnexpress.net/thieu-nu-17-tuoi-thieu-hut-90-hormone-noi-tiet-do-u-tuyen-yen-4739593.html',
-        'https://vnexpress.net/dau-hieu-nhan-biet-hoi-chung-cushing-4738977.html',
-        'https://vnexpress.net/4-cach-giup-giam-can-o-tuoi-day-thi-4737260.html',
-        'https://vnexpress.net/dai-thao-nhat-o-tre-em-co-nguy-hiem-4733833.html',
-        'https://vnexpress.net/6-do-uong-giup-phai-dep-tang-cuong-noi-tiet-to-4731613.html',
-        'https://vnexpress.net/8-meo-tang-noi-tiet-to-cho-nam-gioi-4724086.html',
-        'https://vnexpress.net/lam-dung-thuoc-corticoid-giam-dau-khien-suy-tuyen-thuong-than-4712412.html',
-        'https://vnexpress.net/suy-tuyen-thuong-than-dieu-tri-tai-nha-can-luu-y-gi-4693957.html',
-        'https://vnexpress.net/moi-nguy-do-lam-dung-thuoc-giam-dau-chua-corticoid-4680631.html',
-        'https://vnexpress.net/thua-canxi-trong-mau-co-nguy-hiem-4678388.html',
-        'https://vnexpress.net/dau-hieu-nhan-biet-tut-canxi-4677473.html',
-        'https://vnexpress.net/tuong-phat-tuong-sau-ket-hon-hoa-u-tuyen-thuong-than-4675301.html',
-        'https://vnexpress.net/sot-xuat-huyet-chong-benh-suy-tuyen-yen-4673655.html',
-        'https://vnexpress.net/het-dau-dau-sau-mo-u-thuong-than-4670147.html',
-        'https://vnexpress.net/bien-chung-sau-ba-thang-uong-thuc-pham-chuc-nang-chua-corticoid-4667402.html',
-        'https://vnexpress.net/tai-sao-tre-mac-hoi-chung-cushing-4663890.html',
-        'https://vnexpress.net/estrogen-tac-dong-the-nao-voi-sinh-ly-phai-dep-4660709.html',
-        'https://vnexpress.net/15-thuc-pham-bo-sung-noi-tiet-to-cho-phu-nu-4655321.html',
-        'https://vnexpress.net/dot-quy-tuyen-yen-la-benh-gi-4648556.html',
-        'https://vnexpress.net/dai-thao-nhat-la-benh-gi-4639084.html',
-        'https://vnexpress.net/an-nhieu-duong-tinh-bot-co-anh-huong-sinh-ly-nam-4636974.html',
-        'https://vnexpress.net/16-dau-hieu-can-bo-sung-noi-tiet-to-nu-4636172.html',
-        'https://vnexpress.net/5-cach-bao-ve-da-khoi-mun-noi-tiet-4629030.html',
-        'https://vnexpress.net/9-cach-can-bang-noi-tiet-to-sau-sinh-4626722.html',
-        'https://vnexpress.net/roi-loan-noi-tiet-sau-sinh-co-bat-thuong-4625512.html',
-        'https://vnexpress.net/dau-hieu-nao-canh-bao-suy-giam-noi-tiet-to-4624592.html',
-        'https://vnexpress.net/7-xet-nghiem-noi-tiet-to-nu-thuong-gap-4623694.html',
-        'https://vnexpress.net/8-ly-do-gay-suy-giam-noi-tiet-to-nu-4614428.html',
-        'https://vnexpress.net/dau-hieu-nhan-biet-benh-tuyen-giap-4613003.html',
-        'https://vnexpress.net/10-cach-can-bang-noi-tiet-to-cho-phu-nu-4610304.html',
-        'https://vnexpress.net/cham-soc-da-mun-do-thay-doi-noi-tiet-the-nao-4607042.html',
-        'https://vnexpress.net/lam-the-nao-nhan-biet-mat-can-bang-noi-tiet-to-nu-4606055.html',
-        'https://vnexpress.net/10-benh-lien-quan-den-roi-loan-noi-tiet-4600946.html',
-        'https://vnexpress.net/cach-dieu-tri-nam-da-do-thay-doi-noi-tiet-to-4596093.html',
-        'https://vnexpress.net/phan-biet-nam-nang-va-nam-da-do-thay-doi-noi-tiet-4578924.html',
-        'https://vnexpress.net/u-tuyen-tuy-gay-ha-duong-huyet-hon-me-4578304.html',
-        'https://vnexpress.net/4-nguyen-nhan-thuong-gap-lam-thay-doi-noi-tiet-gay-nam-4574708.html',
-        'https://vnexpress.net/roi-loan-noi-tiet-anh-huong-den-viem-da-co-dia-4490189.html',
-        'https://vnexpress.net/hoi-chung-kallmann-gay-mat-khuu-giac-can-tro-day-thi-4487829.html',
-        'https://vnexpress.net/noi-tiet-to-nu-anh-huong-the-nao-den-suc-khoe-4480771.html',
-        'https://vnexpress.net/cac-benh-roi-loan-noi-tiet-thuong-gap-4479127.html',
-        'https://vnexpress.net/dau-hieu-nhan-biet-phu-nu-beo-bung-do-roi-loan-noi-tiet-4473844.html',
-        'https://vnexpress.net/4-lam-tuong-ve-bo-sung-noi-tiet-to-4454353.html',
-        'https://vnexpress.net/5-cach-can-bang-noi-tiet-to-de-tang-kha-nang-thu-thai-4438878.html',
-        'https://vnexpress.net/nguyen-nhan-gay-roi-loan-noi-tiet-to-kho-thu-thai-4437365.html',
-        'https://vnexpress.net/lieu-phap-noi-tiet-chua-khoi-ung-thu-vu-ma-khong-can-hoa-tri-3786452.html',
-'https://vnexpress.net/6-bien-chung-ung-thu-phoi-4683906.html',
-        'https://vnexpress.net/tai-sao-phu-nu-co-nguy-co-mac-ung-thu-phoi-cao-hon-nam-gioi-4462136.html',
-        'https://vnexpress.net/cach-nhan-biet-cac-truong-hop-khan-cap-ve-ung-thu-phoi-4453067.html',
-        'https://vnexpress.net/9-dau-hieu-canh-bao-benh-phoi-can-som-tam-soat-4449693.html',
-        'https://vnexpress.net/6-dau-hieu-canh-bao-som-benh-ung-thu-phoi-4449179.html',
-        'https://vnexpress.net/phat-hien-som-benh-ho-hap-nho-noi-soi-phe-quan-ong-mem-4444676.html',
-        'https://vnexpress.net/u-phoi-lanh-tinh-phat-hien-som-se-giam-nguy-co-bien-chung-4436932.html',
-        'https://vnexpress.net/hen-o-tre-nho-kho-chan-doan-4391014.html',
-        'https://vnexpress.net/phan-biet-trieu-chung-covid-19-va-benh-phoi-tac-nghen-man-tinh-4381646.html',
-        'https://vnexpress.net/nhung-dieu-can-biet-ve-noi-soi-phe-quan-ong-mem-4351887.html',
-        'https://vnexpress.net/phat-hien-cham-soc-nguoi-benh-phoi-tac-nghen-man-tinh-trong-dich-covid-19-4355556.html'
-'https://vnexpress.net/bat-ngo-phat-hien-u-thuc-quan-khi-noi-soi-da-day-4744532.html',
-        'https://vnexpress.net/beo-phi-thu-pham-gay-nhieu-benh-4705379.html',
-        'https://vnexpress.net/nang-ruot-doi-khien-nguoi-dan-ong-non-oi-suot-hai-nam-4701594.html',
-        'https://vnexpress.net/noi-soi-tieu-hoa-phat-hien-som-dieu-tri-khoi-ung-thu-4693418.html',
-        'https://vnexpress.net/noi-soi-qua-duong-mieng-cat-u-thuc-quan-4680686.html',
-        'https://vnexpress.net/thung-da-day-do-vien-thuoc-nguyen-vi-4676641.html',
-        'https://vnexpress.net/tuong-bat-thuong-thai-ky-hoa-viem-ruot-thua-4663347.html',
-        'https://vnexpress.net/cat-polyp-dai-trang-co-nguy-hiem-khong-4630081.html',
-        'https://vnexpress.net/cu-ong-81-tuoi-chien-thang-hai-benh-ung-thu-4760404.html',
-        'https://vnexpress.net/thuc-pham-nen-an-truoc-khi-noi-soi-dai-trang-4616098.html',
-        'https://vnexpress.net/6-ly-do-can-noi-soi-tieu-hoa-4614443.html',
-        'https://vnexpress.net/noi-soi-duong-mieng-cat-khoi-u-ac-tinh-thuc-quan-4602421.html',
-        'https://vnexpress.net/7-benh-tieu-hoa-co-the-phat-hien-qua-noi-soi-4546028.html',
-        'https://vnexpress.net/nhung-nguoi-nen-thuc-hien-noi-soi-tieu-hoa-4496739.html',
-        'https://vnexpress.net/dieu-tri-ung-thu-gan-bang-dot-song-cao-tan-4483614.html',
-        'https://vnexpress.net/kich-hoat-he-mien-dich-tieu-diet-te-bao-ung-thu-gan-4483243.html',
-        'https://vnexpress.net/truyen-hoa-chat-dong-mach-gan-dieu-tri-ung-thu-gan-tien-trien-4474960.html',
-        'https://vnexpress.net/noi-soi-loai-bo-khoi-u-truc-trang-khong-can-phau-thuat-4473345.html',
-        'https://vnexpress.net/nut-mach-chua-ung-thu-gan-khi-khong-the-phau-thuat-4455023.html',
-        'https://vnexpress.net/chua-tri-ro-hau-mon-phuc-tap-tai-phat-4453897.html',
-        'https://vnexpress.net/mo-noi-soi-u-co-tron-phuc-tap-bang-robot-co-hoc-4452387.html',
-        'https://vnexpress.net/nhung-dieu-can-biet-ve-phau-thuat-cat-ruot-thua-4445044.html',
-        'https://vnexpress.net/can-chuan-bi-gi-khi-cat-dai-trang-4443101.html',
-        'https://vnexpress.net/nhung-dieu-can-biet-ve-phau-thuat-giam-can-4437567.html',
-        'https://vnexpress.net/nhung-luu-y-truoc-khi-thuc-hien-noi-soi-tieu-hoa-4392931.html',
-        'https://vnexpress.net/lan-dau-tien-viet-nam-ung-dung-robot-cam-tay-co-hoc-vao-phau-thuat-noi-soi-4392506.html',
-        'https://vnexpress.net/mo-noi-soi-cho-benh-nhan-bi-thoat-vi-thanh-bung-hiem-gap-4437169.html',
-        'https://vnexpress.net/tiem-xo-noi-soi-giai-phap-chua-tri-khong-phau-thuat-4376815.html',
-        'https://vnexpress.net/nhung-cau-hoi-thuong-gap-ve-phau-thuat-cat-da-day-4353861.html'
-"https://vnexpress.net/xoa-hinh-xam-nhieu-mau-bang-loai-laser-nao-4789281.html",
-        "https://vnexpress.net/5-cach-ngan-seo-loi-phat-trien-4788890.html",
-        "https://vnexpress.net/peel-da-bao-lau-la-du-4788041.html",
-        "https://vnexpress.net/rui-ro-khi-tiem-filler-sai-cach-4787655.html",
-        "https://vnexpress.net/luu-y-truoc-khi-triet-long-vinh-vien-4787438.html",
-        "https://vnexpress.net/nhung-vi-tri-tren-mat-co-the-tiem-filler-4786504.html",
-        "https://vnexpress.net/5-cach-thu-nho-lo-chan-long-4782138.html",
-        "https://vnexpress.net/nhung-tinh-chat-dien-di-lam-dep-pho-bien-4780597.html",
-        "https://vnexpress.net/xoa-nep-nhan-tran-bang-botox-the-nao-4780152.html",
-        "https://vnexpress.net/phuong-phap-lam-dep-voi-axit-tca-4779927.html",
-        "https://vnexpress.net/duong-chat-tu-nhien-giup-giam-ran-da-4778643.html",
-        "https://vnexpress.net/dien-di-tre-hoa-da-nhu-the-nao-4776131.html",
-        "https://vnexpress.net/cach-giam-bong-tham-mat-4775184.html",
-        "https://vnexpress.net/5-tac-dung-lan-kim-voi-da-va-toc-4774647.html",
-        "https://vnexpress.net/cach-cham-soc-vung-da-co-nguc-4772893.html",
-        "https://vnexpress.net/triet-long-co-hieu-qua-bao-lau-4772368.html",
-        "https://vnexpress.net/co-nen-truyen-chat-trang-da-4772312.html",
-        "https://vnexpress.net/phuong-phap-xoa-nong-cam-doi-4771734.html",
-        "https://vnexpress.net/tai-sao-nam-noi-tiet-kho-dieu-tri-4769157.html",
-        "https://vnexpress.net/nhung-dieu-can-biet-ve-tiem-meso-4768668.html",
-        "https://vnexpress.net/phuong-phap-nao-dieu-tri-seo-4767428.html",
-        "https://vnexpress.net/co-nen-dap-mat-na-qua-dem-4767056.html",
-        "https://vnexpress.net/tiem-botox-dieu-tri-cuoi-ho-loi-nhu-the-nao-4765960.html",
-        "https://vnexpress.net/thac-mac-thuong-gap-khi-xoa-hinh-xam-4765584.html",
-        "https://vnexpress.net/dau-hieu-lao-hoa-da-de-nhan-biet-4765259.html",
-        "https://vnexpress.net/dung-tia-laser-xoa-not-ruoi-mi-mat-4764677.html",
-        "https://vnexpress.net/dap-mat-bang-vo-chuoi-co-giup-trang-da-4763979.html",
-        "https://vnexpress.net/lam-gi-khi-tiem-filler-bi-von-cuc-4763975.html",
-        "https://vnexpress.net/tri-lieu-da-bang-cong-nghe-cao-hut-khach-dip-he-4762870.html",
-        "https://vnexpress.net/tai-sao-tiem-filler-o-ma-bi-von-cuc-4761267.html",
-        "https://vnexpress.net/cac-loai-seo-thuong-gap-4759267.html",
-        "https://vnexpress.net/nen-triet-long-vung-kin-bang-phuong-phap-nao-4758496.html",
-        "https://vnexpress.net/tiem-filler-lam-dep-duoc-bao-lau-4756986.html",
-        "https://vnexpress.net/xoa-xam-co-de-lai-seo-4756059.html",
-        "https://vnexpress.net/co-nen-xong-hoi-mat-bang-nuoc-nong-moi-ngay-4755865.html",
-        "https://vnexpress.net/dot-not-ruoi-bang-laser-ngan-hoa-ung-thu-4755507.html",
-        "https://vnexpress.net/co-nen-tu-peel-da-nach-4755046.html",
-        "https://vnexpress.net/thu-pham-gay-nam-da-4754619.html",
-        "https://vnexpress.net/cach-cham-soc-vet-thuong-tranh-seo-4754131.html",
-        "https://vnexpress.net/song-sieu-am-co-giup-giam-beo-4752272.html",
-        "https://vnexpress.net/nen-triet-long-hay-tri-hoi-nach-truoc-4751870.html",
-        "https://vnexpress.net/bo-sung-collagen-cho-da-dung-cach-4751511.html",
-        "https://vnexpress.net/loi-ich-cua-triet-long-4750045.html",
-        "https://vnexpress.net/tay-te-bao-da-chet-bang-ba-ca-phe-4749140.html",
-        "https://vnexpress.net/not-ruoi-sam-mau-ngua-co-phai-ung-thu-4747299.html",
-        "https://vnexpress.net/xoa-bot-mau-ca-phe-sua-bang-laser-4741973.html",
-        "https://vnexpress.net/cach-dieu-tri-tham-moi-4740477.html",
-        "https://vnexpress.net/tai-sao-bi-bien-chung-sau-tiem-filler-4740141.html",
-        "https://vnexpress.net/triet-long-co-het-hoi-nach-4736488.html",
-        "https://vnexpress.net/tiem-botox-lam-dep-co-gay-doc-4732246.html",
-        "https://vnexpress.net/tai-sao-tre-moi-sinh-co-vet-bot-4731737.html",
-        "https://vnexpress.net/tac-dung-cua-tiem-ha-4731192.html",
-        "https://vnexpress.net/tiem-botox-xoa-nhan-the-nao-4730559.html",
-        "https://vnexpress.net/5-cach-tri-seo-ro-4729563.html",
-        "https://vnexpress.net/nhieu-nguoi-triet-long-tri-mui-hoi-mua-nong-4727156.html",
-        "https://vnexpress.net/cach-lam-thon-gon-ham-khong-can-phau-thuat-4721231.html",
-        "https://vnexpress.net/ba-phuong-phap-nhanh-chong-tri-nam-da-4719059.html",
-        "https://vnexpress.net/5-thuc-uong-chong-nang-cho-da-4718210.html",
-        "https://vnexpress.net/ngu-tu-the-nao-de-co-lan-da-dep-4717648.html",
-        "https://vnexpress.net/4-cong-thuc-mat-na-mo-tham-4717517.html",
-        "https://vnexpress.net/an-gi-bo-sung-collagen-cho-da-4717114.html",
-        "https://vnexpress.net/5-loai-mat-na-giam-nep-nhan-4714342.html",
-        "https://vnexpress.net/triet-long-co-hieu-qua-vinh-vien-khong-4714122.html",
-    ]
-                          
-    def parse(self, response):
-        item = VnexpressDataWarehouseItem()
-        #'article_id': response.css('meta[property="og:url"]::attr(content)').get(default='N/A').split('/')[-1],
-        item['title'] = response.css('h1.title-detail::text').get(default='N/A')
-        item['description'] = response.css('div.sidebar-1 > p.description::text').get(default='N/A')
-        item['content'] = ''.join(response.css('article.fck_detail p::text').getall())
-        item['author'] = response.css('div.sidebar-1 > article.fck_detail  > p.Normal:last-child > strong::text').get(default='N/A')
-        item['date'] = response.css('span.date::text').get(default='N/A')
-        item['location'] = response.css('span.location-stamp::text').get()
-        item['disease_name'] = response.css('body > section.section.page-detail.top-detail > div > div.sidebar-1 > div.header-content.width_common > ul > li:nth-child(3) > a::text').get()
-        item['url'] = response.url
-        #'count_comments':response.css('label#total_comment::text').get()
-        yield item
-
-        # Follow the next page link (if exists)
-        next_page = response.css('.next a::attr(href)').get()
-        if next_page:
-            yield response.follow(next_page, self.parse)
